Search SAFE projects by spatial sampling area/named location.

Search for SAFE datasets on the Zenodo database using either a user-provided geometry or the geometry of a named location from the SAFE gazetteer (https://www.safeproject.net/info/gazetteer). The sampling locations provided in each SAFE dataset are tested to see if they intersect the search geometry. A buffer distance can aso be provided to search around the query geometry.

searchSpatial(wkt = NULL, location = NULL, distance = NULL,
  ids = NULL, mostRecent = FALSE)

Arguments

wkt
Character string specifying a well-known text geometry. This is assumed to use latitude and longitude in WGS84 (EPSG:4326), more information on which can be read at https://en.wikipedia.org/wiki/World_Geodetic_System. Useful guidelines on constructing well-known text representations of geometry can be found at https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry.

location
Character string giving a location name used to select query geometry from the SAFE gazetteer, more information on which can be found at https://www.safeproject.net/info/gazetteer.

distance
Numeric search distance, specified in metres. All geometries are converted to UTM 50N projection to provide appropriate distance searching.

ids
Numeric that restricts the returned API search to a subset of Zenodo record IDs given in the ids parameter. This is typically used to refine previous searches.

mostRecent
Logical indicating whether to restrict the API to returning only the most recent versions of the datasets found. By default all versions of the discovered projects are returned. #'

Value

An array of Zenodo record IDs for the SAFE project datasets found to match the search criteria.

Details

searchSpatial will not retrieve datasets that have not provided sampling locations or use newly defined locations that are missing coordinate information. When searching by bounding box (geometric shape) the dataset geographic extent - which is provided for all datasets - is used.

Note that only one of wkt (well-known text geometry) or location (named location on the SAFE gazetteer) should be provided.

Note

This function returns Zenodo record IDs. On the Zenodo database each version of a SAFE project has a unique record ID. Different versions of the same dataset are grouped together under a single concept ID. See https://help.zenodo.org/ for more information on Zenodo versioning. The function will not retrieve datasets for which no location information is provided.

See also

Examples

  searchSpatial(wkt = 'Point(116.5 4.75)')

#> Error in searchSpatial(wkt = "Point(116.5 4.75)"): could not find function "searchSpatial"

  searchSpatial(wkt = 'Point(116.5 4.75)', distance = 100000)

#> Error in searchSpatial(wkt = "Point(116.5 4.75)", distance = 1e+05): could not find function "searchSpatial"

  searchSpatial(wkt = 'Polygon((110 0, 110 10,120 10,120 0,110 0))')

#> Error in searchSpatial(wkt = "Polygon((110 0, 110 10,120 10,120 0,110 0))"): could not find function "searchSpatial"

  searchSpatial(location = 'A_1')

#> Error in searchSpatial(location = "A_1"): could not find function "searchSpatial"

  searchSpatial(location = 'A_1', distance = 2500)

#> Error in searchSpatial(location = "A_1", distance = 2500): could not find function "searchSpatial"

  searchSpatial(location = 'A_1', mostRecent = TRUE)

#> Error in searchSpatial(location = "A_1", mostRecent = TRUE): could not find function "searchSpatial"

  searchSpatial(location = 'A_1', distance = 2500,
                ids = c(1303018, 1237730, 1198585, 1228188, 1198461, 119832))

#> Error in searchSpatial(location = "A_1", distance = 2500, ids = c(1303018,     1237730, 1198585, 1228188, 1198461, 119832)): could not find function "searchSpatial"