SAFE dataset search functions.

In addition to the datasets stored on Zenodo, the SAFE Project website provides an API to search dataset metadata in more depth. The search functions access this API and return safe_record_set objects identifying datasets that match a particular query.

search_dates(dates, match_type = "intersect", most_recent = FALSE, ids = NULL)

search_fields(
  field_text = NULL,
  field_type = NULL,
  ids = NULL,
  most_recent = FALSE
)

search_authors(author, ids = NULL, most_recent = FALSE)

search_taxa(
  taxon_name = NULL,
  taxon_rank = NULL,
  gbif_id = NULL,
  ids = NULL,
  most_recent = FALSE
)

search_text(text, ids = NULL, most_recent = FALSE)

search_spatial(
  wkt = NULL,
  location = NULL,
  distance = NULL,
  ids = NULL,
  most_recent = FALSE
)

Arguments

dates
A vector of length 1 or 2, containing either ISO format date character strings ("yyyy-mm-dd") or POSIXt dates.

match_type
A character string (see Details).

most_recent
Logical indicating whether to restrict the API to returning only the most recent versions of the datasets found. By default all versions of matching dataset concepts are returned.

ids
A set of SAFE dataset record IDs to restrict a search. This will typically be a safe_record_set object returned by another search but can also be a vector of record ids in any of the formats accepted by validate_record_ids.

field_text
Text to search for within the data worksheet field name and description.

field_type
A data worksheet field type (see Links).

author
A character string used to search for datasets by author full (or partial) names.

taxon_name
The scientific name of a taxon to search for.

taxon_rank
A taxonomic rank to search for.

gbif_id
A GBIF taxonomic ID number.

text
Character string to look for within a SAFE dataset, worksheet, title, field description, and dataset keywords.

wkt
A well-known text geometry string, assumed to use latitude and longitude in WGS84 (EPSG:4326).

location
The name of a location in the SAFE gazetteer.

distance
A buffer distance for spatial searches, giving the distance in metres within which to match either location or wkt searches.

Value

An object of class safe_record_set of datasets that match the query.

Details

The API provides endpoints to search datasets by date extents, data worksheet fields, authors, taxa, free text and by spatial query. All of the functions accept the argument most_recent, which restricts the returned datasets to the most recent versions of each matching dataset concept. The functions can also be passed an existing safe_record_set object to search within the results of a previous search.

The match_type parameter specifies how to match date ranges and must be one of "intersect" (default), "contain", or "within". The "contain" option returns datasets that span a date range, "within" returns datasets that fall within the given range and "intersect" selects datasets that overlap any part of the date range. Note that match_type is ignored when only a single date is provided.

Functions

search_dates: Search datasets by date extent

search_fields: Search data worksheet field metadata.

search_authors: Search by dataset author

search_taxa: Search by taxon name, rank or GBIF ID.

search_text: Search dataset, worksheet and field titles and descriptions

search_spatial: Search by spatial sampling area/named location.

Spatial searches

For spatial searches, users can select a location name from a SAFE data gazetteer (see e.g. https://www.safeproject.net/info/gazetteer or load_gazetteer) or provide a WKT geometry. The sampling locations provided in each SAFE dataset are tested to see if they intersect the search geometry.

A buffer distance can aso be provided to extend the search around the query geometry. Note that although WKT geometries should be provided using WGS84 lat/long coordinates, since this is typical field GPS data, distances must be provided as metres and all proximity calculations take place in the UTM50N projected coordinate system.

The search_spatial function will not retrieve datasets that have not provided sampling locations or use newly defined locations that are missing coordinate information.

Links

SAFE data API

e.g. https://www.safeproject.net/api

Worksheet field types

https://safedata-validator.readthedocs.io/en/latest/data_format/data/#field-types

SAFE gazetteer

See load_gazetteer and e.g. https://www.safeproject.net/info/gazetteer

WKT

https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry

Examples

   # \donttest{
   search_dates("2014-06-12")

#> Search returned 59 records

#> Set includes 46 concept ids and 59 record ids: 
#>  - 38 open and most recent (*)
#>  - 12 open and outdated (o)
#>  - 9 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   629855 3490675         *
#> 2   629942 3492153         *
#> 3   630004 3490488         *
#> 4   630015 3492009         *
#> 5   630063 3697796         *
#> 6   630119 3492148         *
#> 7   630123 3491956         *
#> 8   630132 3490613         *
#> 9  ------- 3490564         o
#> 10  630161 3492064         *
#> 11  630228 3697804         *
#> 12  630241 3492126         *
#> 13  630259 3490687         *
#> 14  654585 3698297         *
#> 15 1198465 1237733         *
#> 16 ------- 1198466         o
#> 17 1198468 1237731         *
#> 18 ------- 1198469         o
#> 19 1198692 2599732         *
#> 20 ------- 1198693         o
#> 21 1198696 1198846         *
#> 22 ------- 1198697         o
#> 23 1198840 1237719         *
#> 24 ------- 1198841         o
#> 25 1216002 1304118         *
#> 26 ------- 1237657         o
#> 27 ------- 1216003         o
#> 28 1303011 1303012         *
#> 29 1303013 1303014         *
#> 30 1327556 1327557         *
#> 31 1344568 1344569         *
#> 32 1400563 1400564         *
#> 33 1441584 1441585         *
#> 34 1478523 3955050         x
#> 35 ------- 1478524         x
#> 36 1478525 1478526         *
#> 37 1995247 1995439         *
#> 38 ------- 1995248         o
#> 39 3247601 3247602         x
#> 40 3247630 3247631         x
#> 41 3247638 3247639         *
#> 42 3266770 3266771         x
#> 43 3266826 3266827         *
#> 44 3342494 3342495         *
#> 45 3354067 3354068         *
#> 46 3366103 3366104         *
#> 47 3402745 3402746         x
#> 48 3402747 3402748         x
#> 49 3402751 3402752         x
#> 50 3476357 3476390         *
#> 51 ------- 3476358         o
#> 52 3485120 3906082         *
#> 53 ------- 3833856         o
#> 54 ------- 3485121         o
#> 55 3780390 3780391         x
#> 56 3888374 3888375         *
#> 57 3901504 3901505         *
#> 58 3908248 3908249         *
#> 59 3948442 3948443         *

   search_dates(as.POSIXct(c("2014-06-12", "2015-06-11")))

#> Search returned 79 records

#> Set includes 62 concept ids and 79 record ids: 
#>  - 50 open and most recent (*)
#>  - 16 open and outdated (o)
#>  - 13 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   629855 3490675         *
#> 2   629942 3492153         *
#> 3   630004 3490488         *
#> 4   630015 3492009         *
#> 5   630063 3697796         *
#> 6   630119 3492148         *
#> 7   630123 3491956         *
#> 8   630132 3490613         *
#> 9  ------- 3490564         o
#> 10  630161 3492064         *
#> 11  630228 3697804         *
#> 12  630241 3492126         *
#> 13  630259 3490687         *
#> 14  654585 3698297         *
#> 15 1198465 1237733         *
#> 16 ------- 1198466         o
#> 17 1198468 1237731         *
#> 18 ------- 1198469         o
#> 19 1198471 1237732         *
#> 20 ------- 1198472         o
#> 21 1198692 2599732         *
#> 22 ------- 1198693         o
#> 23 1198696 1198846         *
#> 24 ------- 1198697         o
#> 25 1198840 1237719         *
#> 26 ------- 1198841         o
#> 27 1216002 1304118         *
#> 28 ------- 1237657         o
#> 29 ------- 1216003         o
#> 30 1303011 1303012         *
#> 31 1303013 1303014         *
#> 32 1303019 1303020         *
#> 33 1327556 1327557         *
#> 34 1344568 1344569         *
#> 35 1400563 1400564         *
#> 36 1441584 1441585         *
#> 37 1478523 3955050         x
#> 38 ------- 1478524         x
#> 39 1478525 1478526         *
#> 40 1487594 1487595         *
#> 41 1995247 1995439         *
#> 42 ------- 1995248         o
#> 43 2020339 2020340         *
#> 44 2536268 3476542         *
#> 45 ------- 2536269         o
#> 46 3081058 3247699         *
#> 47 ------- 3081059         o
#> 48 3247464 3247465         x
#> 49 3247493 3247494         x
#> 50 3247601 3247602         x
#> 51 3247630 3247631         x
#> 52 3247638 3247639         *
#> 53 3251899 3258117         *
#> 54 ------- 3251900         o
#> 55 3265703 3265704         *
#> 56 3265745 3265746         *
#> 57 3265747 3265748         *
#> 58 3266770 3266771         x
#> 59 3266826 3266827         *
#> 60 3342494 3342495         *
#> 61 3354067 3354068         *
#> 62 3366103 3366104         *
#> 63 3402745 3402746         x
#> 64 3402747 3402748         x
#> 65 3402751 3402752         x
#> 66 3476357 3476390         *
#> 67 ------- 3476358         o
#> 68 3485120 3906082         *
#> 69 ------- 3833856         o
#> 70 ------- 3485121         o
#> 71 3698114 3698115         x
#> 72 3780390 3780391         x
#> 73 3888374 3888375         *
#> 74 3901504 3901505         *
#> 75 3906117 3906118         x
#> 76 3908127 3908128         *
#> 77 3908248 3908249         *
#> 78 3929631 3929632         *
#> 79 3948442 3948443         *

   search_dates(c("2014-06-12", "2015-06-11"), match_type = "contain")

#> Search returned 5 records

#> Set includes 4 concept ids and 5 record ids: 
#>  - 3 open and most recent (*)
#>  - 1 open and outdated (o)
#>  - 1 under embargo or restricted (x)
#> 
#>   concept  record available
#> 1 2020339 2020340         *
#> 2 3081058 3247699         *
#> 3 ------- 3081059         o
#> 4 3247493 3247494         x
#> 5 3265703 3265704         *

   search_fields(field_text = "temperature")

#> Search returned 17 records

#> Set includes 14 concept ids and 17 record ids: 
#>  - 11 open and most recent (*)
#>  - 2 open and outdated (o)
#>  - 4 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  3247484 3876227         *
#> 2  ------- 3247485         o
#> 3  3247591 3247592         *
#> 4  3251886 3258079         *
#> 5  3251899 3258117         *
#> 6  ------- 3251900         o
#> 7  3266770 3266771         x
#> 8  3266821 3266822         *
#> 9  3475407 3475408         *
#> 10 3479170 3479171         *
#> 11 3698114 3698115         x
#> 12 3888374 3888375         *
#> 13 3908146 3929764         x
#> 14 ------- 3908147         x
#> 15 3926373 3926374         *
#> 16 3948442 3948443         *
#> 17 3952831 3952832         *

   search_fields(field_type = "numeric")

#> Search returned 68 records

#> Set includes 56 concept ids and 68 record ids: 
#>  - 40 open and most recent (*)
#>  - 10 open and outdated (o)
#>  - 18 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   629855 3490675         *
#> 2   629942 3492153         *
#> 3   630015 3492009         *
#> 4   630119 3492148         *
#> 5   630132 3490613         *
#> 6   630161 3492064         *
#> 7   630241 3492126         *
#> 8  1198564 3857210         *
#> 9  1478523 3955050         x
#> 10 2536268 3476542         *
#> 11 2537072 3901735         x
#> 12 3081058 3247699         *
#> 13 3237506 3237507         x
#> 14 3247464 3247465         x
#> 15 3247484 3876227         *
#> 16 ------- 3247485         o
#> 17 3247491 3247492         *
#> 18 3247493 3247494         x
#> 19 3247591 3247592         *
#> 20 3247601 3247602         x
#> 21 3247630 3247631         x
#> 22 3247638 3247639         *
#> 23 3251886 3258079         *
#> 24 ------- 3251887         o
#> 25 3251899 3258117         *
#> 26 ------- 3251900         o
#> 27 3251901 3897394         *
#> 28 ------- 3614876         o
#> 29 ------- 3258082         o
#> 30 ------- 3251902         o
#> 31 3265703 3265704         *
#> 32 3265711 3742834         x
#> 33 ------- 3265712         x
#> 34 3265721 3265722         *
#> 35 3265745 3265746         *
#> 36 3265747 3265748         *
#> 37 3266770 3266771         x
#> 38 3266821 3266822         *
#> 39 3342494 3342495         *
#> 40 3354067 3354068         *
#> 41 3366103 3366104         *
#> 42 3402745 3402746         x
#> 43 3475405 3475406         *
#> 44 3475407 3475408         *
#> 45 3476357 3476390         *
#> 46 ------- 3476358         o
#> 47 3479170 3479171         *
#> 48 3485085 3485086         x
#> 49 3485120 3906082         *
#> 50 ------- 3833856         o
#> 51 ------- 3485121         o
#> 52 3698114 3698115         x
#> 53 3780390 3780391         x
#> 54 3888374 3888375         *
#> 55 3897376 3897377         x
#> 56 3901504 3901505         *
#> 57 3906117 3906118         x
#> 58 3908117 3908118         *
#> 59 3908127 3908128         *
#> 60 3908146 3929764         x
#> 61 ------- 3908147         x
#> 62 3908248 3908249         *
#> 63 3926373 3926374         *
#> 64 3929631 3929632         *
#> 65 3948442 3948443         *
#> 66 3952831 3952832         *
#> 67 3966236 3966277         *
#> 68 ------- 3966237         o

   search_fields(field_text = "temperature", field_type = "numeric")

#> Search returned 16 records

#> Set includes 13 concept ids and 16 record ids: 
#>  - 10 open and most recent (*)
#>  - 2 open and outdated (o)
#>  - 4 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  3247484 3876227         *
#> 2  ------- 3247485         o
#> 3  3247591 3247592         *
#> 4  3251886 3258079         *
#> 5  3251899 3258117         *
#> 6  ------- 3251900         o
#> 7  3266770 3266771         x
#> 8  3266821 3266822         *
#> 9  3475407 3475408         *
#> 10 3479170 3479171         *
#> 11 3698114 3698115         x
#> 12 3888374 3888375         *
#> 13 3908146 3929764         x
#> 14 ------- 3908147         x
#> 15 3926373 3926374         *
#> 16 3952831 3952832         *

   search_authors("Ewers")

#> Search returned 72 records

#> Set includes 56 concept ids and 72 record ids: 
#>  - 49 open and most recent (*)
#>  - 13 open and outdated (o)
#>  - 10 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   629855 3490675         *
#> 2   629942 3492153         *
#> 3   654585 3698297         *
#> 4  1198301 1198302         *
#> 5  1198460 1237736         *
#> 6  1198465 1237733         *
#> 7  1198468 1237731         *
#> 8  1198471 1237732         *
#> 9  1198476 1198477         *
#> 10 1198522 1237724         *
#> 11 1198560 1237722         *
#> 12 1198568 1304117         *
#> 13 ------- 1237721         o
#> 14 1198576 1237737         *
#> 15 ------- 1198577         o
#> 16 1198578 1237738         *
#> 17 ------- 1198579         o
#> 18 1198580 1237726         *
#> 19 1198582 1237725         *
#> 20 1198584 1198585         *
#> 21 1198586 1198587         *
#> 22 1198596 1237739         *
#> 23 1198692 2599732         *
#> 24 ------- 1198693         o
#> 25 1198694 1198695         *
#> 26 1198696 1198846         *
#> 27 ------- 1198697         o
#> 28 1198838 1198839         *
#> 29 1198840 1237719         *
#> 30 1198842 1237720         *
#> 31 1216002 1304118         *
#> 32 ------- 1237657         o
#> 33 ------- 1216003         o
#> 34 1216039 2537074         *
#> 35 ------- 1400565         o
#> 36 ------- 1216040         o
#> 37 1216043 1216044         *
#> 38 1219160 1219161         *
#> 39 1220269 1220270         *
#> 40 1228187 1228188         *
#> 41 1228189 1228190         *
#> 42 1237729 1237730         x
#> 43 1303011 1303012         *
#> 44 1303013 1303014         *
#> 45 1303017 1303018         *
#> 46 1303019 1303020         *
#> 47 1323503 1323504         *
#> 48 1344568 1344569         *
#> 49 1400563 1400564         *
#> 50 1441584 1441585         *
#> 51 1478523 3955050         x
#> 52 2537072 3901735         x
#> 53 ------- 2537073         x
#> 54 3081058 3247699         *
#> 55 ------- 3081059         o
#> 56 3237506 3237507         x
#> 57 3247484 3876227         *
#> 58 ------- 3247485         o
#> 59 3265711 3742834         x
#> 60 ------- 3265712         x
#> 61 3265721 3265722         *
#> 62 3266770 3266771         x
#> 63 3354067 3354068         *
#> 64 3476357 3476390         *
#> 65 ------- 3476358         o
#> 66 3479170 3479171         *
#> 67 3888374 3888375         *
#> 68 3908117 3908118         *
#> 69 3908146 3929764         x
#> 70 ------- 3908147         x
#> 71 3966236 3966277         *
#> 72 ------- 3966237         o

   search_taxa(taxon_name = "Formicidae")

#> Search returned 11 records

#> Set includes 10 concept ids and 11 record ids: 
#>  - 8 open and most recent (*)
#>  - 1 open and outdated (o)
#>  - 2 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1198301 1198302         *
#> 2  1198471 1237732         *
#> 3  1198838 1198839         *
#> 4  1237729 1237730         x
#> 5  1400561 1400562         *
#> 6  1995247 1995439         *
#> 7  2537072 3901735         x
#> 8  3247484 3876227         *
#> 9  ------- 3247485         o
#> 10 3265745 3265746         *
#> 11 3354067 3354068         *

   search_taxa(gbif_id = 4342)

#> Search returned 11 records

#> Set includes 10 concept ids and 11 record ids: 
#>  - 8 open and most recent (*)
#>  - 1 open and outdated (o)
#>  - 2 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1198301 1198302         *
#> 2  1198471 1237732         *
#> 3  1198838 1198839         *
#> 4  1237729 1237730         x
#> 5  1400561 1400562         *
#> 6  1995247 1995439         *
#> 7  2537072 3901735         x
#> 8  3247484 3876227         *
#> 9  ------- 3247485         o
#> 10 3265745 3265746         *
#> 11 3354067 3354068         *

   search_taxa(taxon_rank = "family")

#> Search returned 72 records

#> Set includes 66 concept ids and 72 record ids: 
#>  - 49 open and most recent (*)
#>  - 3 open and outdated (o)
#>  - 20 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1198301 1198302         *
#> 2  1198460 1237736         *
#> 3  1198465 1237733         *
#> 4  1198468 1237731         *
#> 5  1198471 1237732         *
#> 6  1198474 1198475         *
#> 7  1198476 1198477         *
#> 8  1198522 1237724         *
#> 9  1198560 1237722         *
#> 10 1198568 1304117         *
#> 11 1198596 1237739         *
#> 12 1198692 2599732         *
#> 13 1198694 1198695         *
#> 14 1198696 1198846         *
#> 15 1198832 1198833         *
#> 16 1198838 1198839         *
#> 17 1198840 1237719         *
#> 18 1198842 1237720         *
#> 19 1216039 2537074         *
#> 20 1219160 1219161         *
#> 21 1237729 1237730         x
#> 22 1303009 1303010         *
#> 23 1303019 1303020         *
#> 24 1323503 1323504         *
#> 25 1327556 1327557         *
#> 26 1400561 1400562         *
#> 27 1400563 1400564         *
#> 28 1478523 3955050         x
#> 29 ------- 1478524         x
#> 30 1478525 1478526         *
#> 31 1487594 1487595         *
#> 32 1995247 1995439         *
#> 33 2020339 2020340         *
#> 34 2024580 2536270         *
#> 35 2536268 3476542         *
#> 36 ------- 2536269         o
#> 37 2537072 3901735         x
#> 38 2579791 2579792         *
#> 39 3237506 3237507         x
#> 40 3247464 3247465         x
#> 41 3247484 3876227         *
#> 42 ------- 3247485         o
#> 43 3247491 3247492         *
#> 44 3247493 3247494         x
#> 45 3247630 3247631         x
#> 46 3265703 3265704         *
#> 47 3265711 3742834         x
#> 48 ------- 3265712         x
#> 49 3265745 3265746         *
#> 50 3342494 3342495         *
#> 51 3354067 3354068         *
#> 52 3366103 3366104         *
#> 53 3402747 3402748         x
#> 54 3402751 3402752         x
#> 55 3475405 3475406         *
#> 56 3475407 3475408         *
#> 57 3485085 3485086         x
#> 58 3780390 3780391         x
#> 59 3832075 3832076         x
#> 60 3897376 3897377         x
#> 61 3901504 3901505         *
#> 62 3906117 3906118         x
#> 63 3906440 3906441         x
#> 64 3908127 3908128         *
#> 65 3908146 3929764         x
#> 66 ------- 3908147         x
#> 67 3908248 3908249         *
#> 68 3926373 3926374         *
#> 69 3948442 3948443         *
#> 70 3952831 3952832         *
#> 71 3966236 3966277         *
#> 72 ------- 3966237         o

   search_text("forest")

#> Search returned 52 records

#> Set includes 43 concept ids and 52 record ids: 
#>  - 28 open and most recent (*)
#>  - 7 open and outdated (o)
#>  - 17 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   629942 3492153         *
#> 2   630132 3490613         *
#> 3   630161 3492064         *
#> 4  1478523 3955050         x
#> 5  2536268 3476542         *
#> 6  2537072 3901735         x
#> 7  3081058 3247699         *
#> 8  3237506 3237507         x
#> 9  3247464 3247465         x
#> 10 3247484 3876227         *
#> 11 ------- 3247485         o
#> 12 3247591 3247592         *
#> 13 3247601 3247602         x
#> 14 3247630 3247631         x
#> 15 3247638 3247639         *
#> 16 3251886 3258079         *
#> 17 ------- 3251887         o
#> 18 3251899 3258117         *
#> 19 ------- 3251900         o
#> 20 3265703 3265704         *
#> 21 3265711 3742834         x
#> 22 ------- 3265712         x
#> 23 3265745 3265746         *
#> 24 3265747 3265748         *
#> 25 3266770 3266771         x
#> 26 3266821 3266822         *
#> 27 3266826 3266827         *
#> 28 3366103 3366104         *
#> 29 3402747 3402748         x
#> 30 3402751 3402752         x
#> 31 3475405 3475406         *
#> 32 3475407 3475408         *
#> 33 3476357 3476390         *
#> 34 ------- 3476358         o
#> 35 3485085 3485086         x
#> 36 3485120 3906082         *
#> 37 ------- 3833856         o
#> 38 ------- 3485121         o
#> 39 3698114 3698115         x
#> 40 3780390 3780391         x
#> 41 3897376 3897377         x
#> 42 3901504 3901505         *
#> 43 3908117 3908118         *
#> 44 3908127 3908128         *
#> 45 3908146 3929764         x
#> 46 ------- 3908147         x
#> 47 3926373 3926374         *
#> 48 3929631 3929632         *
#> 49 3948442 3948443         *
#> 50 3952831 3952832         *
#> 51 3966236 3966277         *
#> 52 ------- 3966237         o

   search_text("ant")

#> Search returned 45 records

#> Set includes 41 concept ids and 45 record ids: 
#>  - 24 open and most recent (*)
#>  - 2 open and outdated (o)
#>  - 19 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   630119 3492148         *
#> 2   630161 3492064         *
#> 3  1478523 3955050         x
#> 4  2536268 3476542         *
#> 5  2537072 3901735         x
#> 6  3081058 3247699         *
#> 7  3247464 3247465         x
#> 8  3247484 3876227         *
#> 9  ------- 3247485         o
#> 10 3247491 3247492         *
#> 11 3247493 3247494         x
#> 12 3247601 3247602         x
#> 13 3247630 3247631         x
#> 14 3247638 3247639         *
#> 15 3265703 3265704         *
#> 16 3265711 3742834         x
#> 17 ------- 3265712         x
#> 18 3265745 3265746         *
#> 19 3265747 3265748         *
#> 20 3266770 3266771         x
#> 21 3266821 3266822         *
#> 22 3266826 3266827         *
#> 23 3354067 3354068         *
#> 24 3402745 3402746         x
#> 25 3402751 3402752         x
#> 26 3475405 3475406         *
#> 27 3475407 3475408         *
#> 28 3485085 3485086         x
#> 29 3780390 3780391         x
#> 30 3832075 3832076         x
#> 31 3888374 3888375         *
#> 32 3897376 3897377         x
#> 33 3901504 3901505         *
#> 34 3906117 3906118         x
#> 35 3906440 3906441         x
#> 36 3908117 3908118         *
#> 37 3908127 3908128         *
#> 38 3908146 3929764         x
#> 39 ------- 3908147         x
#> 40 3926373 3926374         *
#> 41 3929631 3929632         *
#> 42 3948442 3948443         *
#> 43 3952831 3952832         *
#> 44 3966236 3966277         *
#> 45 ------- 3966237         o

   search_spatial(wkt = "Point(116.5 4.75)")

#> Search returned 0 records

#> Set includes 0 concept ids and 0 record ids: 
#>  - 0 open and most recent (*)
#>  - 0 open and outdated (o)
#>  - 0 under embargo or restricted (x)
#> 
#> [1] concept   record    available
#> <0 rows> (or 0-length row.names)

   search_spatial(wkt = "Point(116.5 4.75)", distance = 100000)

#> Search returned 60 records

#> Set includes 48 concept ids and 60 record ids: 
#>  - 37 open and most recent (*)
#>  - 10 open and outdated (o)
#>  - 13 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   630015 3492009         *
#> 2   630132 3490613         *
#> 3  ------- 3490564         o
#> 4  1198460 1237736         *
#> 5  ------- 1198461         o
#> 6  1198476 1198477         *
#> 7  1198564 3857210         *
#> 8  ------- 1198565         o
#> 9  1198576 1237737         *
#> 10 ------- 1198577         o
#> 11 1198578 1237738         *
#> 12 ------- 1198579         o
#> 13 1198580 1237726         *
#> 14 ------- 1198581         o
#> 15 1198582 1237725         *
#> 16 ------- 1198583         o
#> 17 1198584 1198585         *
#> 18 1198586 1198587         *
#> 19 1198692 2599732         *
#> 20 ------- 1198693         o
#> 21 1198696 1198846         *
#> 22 ------- 1198697         o
#> 23 1198832 1198833         *
#> 24 1198838 1198839         *
#> 25 1216043 1216044         *
#> 26 1220269 1220270         *
#> 27 1228187 1228188         *
#> 28 1228189 1228190         *
#> 29 1237729 1237730         x
#> 30 1303017 1303018         *
#> 31 1303019 1303020         *
#> 32 1323503 1323504         *
#> 33 1344568 1344569         *
#> 34 1400563 1400564         *
#> 35 1441584 1441585         *
#> 36 1478523 3955050         x
#> 37 ------- 1478524         x
#> 38 2537072 3901735         x
#> 39 ------- 2537073         x
#> 40 3247484 3876227         *
#> 41 ------- 3247485         o
#> 42 3247491 3247492         *
#> 43 3247493 3247494         x
#> 44 3247591 3247592         *
#> 45 3247601 3247602         x
#> 46 3247630 3247631         x
#> 47 3247638 3247639         *
#> 48 3265721 3265722         *
#> 49 3265745 3265746         *
#> 50 3265747 3265748         *
#> 51 3266770 3266771         x
#> 52 3342494 3342495         *
#> 53 3354067 3354068         *
#> 54 3402747 3402748         x
#> 55 3402751 3402752         x
#> 56 3780390 3780391         x
#> 57 3906440 3906441         x
#> 58 3908117 3908118         *
#> 59 3908127 3908128         *
#> 60 3952831 3952832         *

   search_spatial(wkt = "Polygon((110 0, 110 10,120 10,120 0,110 0))")

#> Search returned 104 records

#> Set includes 79 concept ids and 104 record ids: 
#>  - 62 open and most recent (*)
#>  - 22 open and outdated (o)
#>  - 20 under embargo or restricted (x)
#> 
#>     concept  record available
#> 1    629855 3490675         *
#> 2    630015 3492009         *
#> 3    630132 3490613         *
#> 4   ------- 3490564         o
#> 5   1198460 1237736         *
#> 6   ------- 1198461         o
#> 7   1198465 1237733         *
#> 8   ------- 1198466         o
#> 9   1198468 1237731         *
#> 10  ------- 1198469         o
#> 11  1198471 1237732         *
#> 12  ------- 1198472         o
#> 13  1198474 1198475         *
#> 14  1198476 1198477         *
#> 15  1198556 1198557         *
#> 16  1198560 1237722         *
#> 17  ------- 1198561         o
#> 18  1198564 3857210         *
#> 19  ------- 1198565         o
#> 20  1198576 1237737         *
#> 21  ------- 1198577         o
#> 22  1198578 1237738         *
#> 23  ------- 1198579         o
#> 24  1198580 1237726         *
#> 25  ------- 1198581         o
#> 26  1198582 1237725         *
#> 27  ------- 1198583         o
#> 28  1198584 1198585         *
#> 29  1198586 1198587         *
#> 30  1198596 1237739         *
#> 31  ------- 1198597         o
#> 32  1198692 2599732         *
#> 33  ------- 1198693         o
#> 34  1198696 1198846         *
#> 35  ------- 1198697         o
#> 36  1198832 1198833         *
#> 37  1198838 1198839         *
#> 38  1198842 1237720         *
#> 39  ------- 1198843         o
#> 40  1216043 1216044         *
#> 41  1219160 1219161         *
#> 42  1220269 1220270         *
#> 43  1228187 1228188         *
#> 44  1228189 1228190         *
#> 45  1237729 1237730         x
#> 46  1303009 1303010         *
#> 47  1303017 1303018         *
#> 48  1303019 1303020         *
#> 49  1323503 1323504         *
#> 50  1327556 1327557         *
#> 51  1344568 1344569         *
#> 52  1400563 1400564         *
#> 53  1441584 1441585         *
#> 54  1478523 3955050         x
#> 55  ------- 1478524         x
#> 56  1478525 1478526         *
#> 57  1995247 1995439         *
#> 58  ------- 1995248         o
#> 59  2024580 2536270         *
#> 60  ------- 2024581         o
#> 61  2536268 3476542         *
#> 62  ------- 2536269         o
#> 63  2537072 3901735         x
#> 64  ------- 2537073         x
#> 65  3081058 3247699         *
#> 66  ------- 3081059         o
#> 67  3237506 3237507         x
#> 68  3247484 3876227         *
#> 69  ------- 3247485         o
#> 70  3247491 3247492         *
#> 71  3247493 3247494         x
#> 72  3247591 3247592         *
#> 73  3247601 3247602         x
#> 74  3247630 3247631         x
#> 75  3247638 3247639         *
#> 76  3265721 3265722         *
#> 77  3265745 3265746         *
#> 78  3265747 3265748         *
#> 79  3266770 3266771         x
#> 80  3342494 3342495         *
#> 81  3354067 3354068         *
#> 82  3402747 3402748         x
#> 83  3402751 3402752         x
#> 84  3475407 3475408         *
#> 85  3476357 3476390         *
#> 86  ------- 3476358         o
#> 87  3479170 3479171         *
#> 88  3485085 3485086         x
#> 89  3780390 3780391         x
#> 90  3832075 3832076         x
#> 91  3897376 3897377         x
#> 92  3901504 3901505         *
#> 93  3906117 3906118         x
#> 94  3906440 3906441         x
#> 95  3908117 3908118         *
#> 96  3908127 3908128         *
#> 97  3908146 3929764         x
#> 98  ------- 3908147         x
#> 99  3926373 3926374         *
#> 100 3929631 3929632         *
#> 101 3948442 3948443         *
#> 102 3952831 3952832         *
#> 103 3966236 3966277         *
#> 104 ------- 3966237         o

   search_spatial(location = "A_1")

#> Search returned 9 records

#> Set includes 7 concept ids and 9 record ids: 
#>  - 7 open and most recent (*)
#>  - 2 open and outdated (o)
#>  - 0 under embargo or restricted (x)
#> 
#>   concept  record available
#> 1 1198460 1237736         *
#> 2 ------- 1198461         o
#> 3 1198476 1198477         *
#> 4 1228187 1228188         *
#> 5 1323503 1323504         *
#> 6 1344568 1344569         *
#> 7 3354067 3354068         *
#> 8 3966236 3966277         *
#> 9 ------- 3966237         o

   search_spatial(location = "A_1", distance = 2500)

#> Search returned 45 records

#> Set includes 35 concept ids and 45 record ids: 
#>  - 29 open and most recent (*)
#>  - 10 open and outdated (o)
#>  - 6 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   629855 3490675         *
#> 2  1198460 1237736         *
#> 3  ------- 1198461         o
#> 4  1198468 1237731         *
#> 5  ------- 1198469         o
#> 6  1198476 1198477         *
#> 7  1198576 1237737         *
#> 8  ------- 1198577         o
#> 9  1198578 1237738         *
#> 10 ------- 1198579         o
#> 11 1198584 1198585         *
#> 12 1198586 1198587         *
#> 13 1228187 1228188         *
#> 14 1228189 1228190         *
#> 15 1237729 1237730         x
#> 16 1303009 1303010         *
#> 17 1303017 1303018         *
#> 18 1323503 1323504         *
#> 19 1344568 1344569         *
#> 20 1400563 1400564         *
#> 21 1441584 1441585         *
#> 22 1478525 1478526         *
#> 23 1995247 1995439         *
#> 24 ------- 1995248         o
#> 25 2024580 2536270         *
#> 26 ------- 2024581         o
#> 27 2536268 3476542         *
#> 28 ------- 2536269         o
#> 29 3247484 3876227         *
#> 30 ------- 3247485         o
#> 31 3247491 3247492         *
#> 32 3247493 3247494         x
#> 33 3265721 3265722         *
#> 34 3354067 3354068         *
#> 35 3476357 3476390         *
#> 36 ------- 3476358         o
#> 37 3485085 3485086         x
#> 38 3832075 3832076         x
#> 39 3897376 3897377         x
#> 40 3901504 3901505         *
#> 41 3906117 3906118         x
#> 42 3908127 3908128         *
#> 43 3926373 3926374         *
#> 44 3966236 3966277         *
#> 45 ------- 3966237         o

   # }