Update the SAFE index file

The SAFE index file contains version and file-download information for all SAFE projects that have been requested by the user through the safedata package. The intended use of the index file is to provide an offline cache of information associated with the varied records on the SAFE Zenodo database to streamline operation of safedata and reduce redundancies. The index file is a dataframe stored (by default) in the SAFE_data_dir. Records are grouped by concept ID. Meta information, including DOI, record ID, date created, ddataset embargo status, open date, and local file download status (TRUE/FALSE) are then saved for each record within the concept IDs requested. Every time a call is made to the Zenodo database for a SAFE dataset, updateIndex runs, either adding new version data or overwriting existing entries, depending on whether the concept ID has been previously requested.

updateIndex(id, isConceptId = TRUE, dir = NULL)

Arguments

id
the concept ID/record ID associated with the version of the SAFE project being accessed

isConceptId
boolean indicating whether parameter ID is a concept ID or a record ID. This changes the operation of the function

dir
the directory into which the index.rds file should be stored, which defaults to the SAFE_data_dir

Details

updateIndex takes a Zenodo ID as its main input, which may be a concept ID (the default, isConceptId = TRUE), or the ID of a specific record entry (isConceptId = FALSE). Under the default behaviour, information pertaining to all versions of a particular concept are accessed through a call to zenodoVersionsApiLookup and stored within the index file. When a record ID is passed as input, it is assumed that it is because the dataset associated with this version has been downloaded and, after a directory check, the download status of this entry is modified. At the end of each call the modified index dataframe is saved in the local SAFE directory.

The SAFE index file is used downstream to identify and notify the user of the most recent available versions of a project, or in the event that the user is attempting to access a dataset that already exists within the SAFE_data_dir.

See also

Examples

  # add version information for concept ID 3081058
  setSafeDir("C:/Users/User/SAFE_data/")

#> Error in setSafeDir("C:/Users/User/SAFE_data/"): could not find function "setSafeDir"

  updateIndex(3081058)

#> Error in updateIndex(3081058): Directory not supplied and SAFE_data_dir not set! See setSafeDir() function

  # update download status of record 3081059 to TRUE
  updateIndex(3081059, isConceptId = FALSE)

#> Error in updateIndex(3081059, isConceptId = FALSE): Directory not supplied and SAFE_data_dir not set! See setSafeDir() function