Download SAFE dataset files

This downloads files associated with SAFE datasets, either all of the files included in a set of records (xlsx_only=FALSE) or just the core .xlsx files (xlsx_only=FALSE), and stores them in the SAFE data directory. Currently, there is no mechanism for importing restricted datasets.

download_safe_files(
  record_ids,
  confirm = TRUE,
  xlsx_only = TRUE,
  download_metadata = TRUE,
  refresh = FALSE,
  token = NULL
)

Arguments

record_ids
A vector of SAFE dataset record ids or a safe_record_set object.

confirm
Requires the user to confirm before download (logical)

xlsx_only
Should all files be downloaded or just the core .xslx file (logical)

download_metadata
Should the metadata record for the file be downloaded (logical)

refresh
Should the function check if local copies have been modified and download fresh copies. This is useful if the local copies have unintentionally been modified but note the warning above.

token
An access token for restricted datasets. Not currently implemented.

Value

Invisibly, a vector of paths for successfully downloaded files.

Details

By default, the function will also download the dataset metadata. This information is required by many of the functions in the package but users can turn off automatic metadata download.

Warning

Using refresh=TRUE will overwrite locally modified files and replace them with the versions of record from Zenodo.

Examples

   # \donttest{
   set_example_safe_dir()
   recs <- validate_record_ids(c(3247631, 3266827, 3266821))
   download_safe_files(recs, confirm=FALSE)

#> 2 files requested from 2 records
#>  - 0 local (0 bytes)
#>  - 1 embargoed or restricted (647.1 Kb)
#>  - 1 to download (27.9 Kb)

#> Downloading 2 record metadata files

#> 1 files for record 3247631: under embargo or restricted

#> 1 files for record 3266827: 1 to download

#>  - Downloaded: template_Symes.xlsx

   unset_example_safe_dir()
   # }