SAFE dataset search functions.

In addition to the datasets stored on Zenodo, the SAFE Project website provides an API to search dataset metadata in more depth. The search functions access this API and return safe_record_set objects identifying datasets that match a particular query.

search_dates(dates, match_type = "intersect", most_recent = FALSE,
  ids = NULL)

search_fields(field_text = NULL, field_type = NULL, ids = NULL,
  most_recent = FALSE)

search_authors(author, ids = NULL, most_recent = FALSE)

search_taxa(taxon_name = NULL, taxon_rank = NULL, gbif_id = NULL,
  ids = NULL, most_recent = FALSE)

search_text(text, ids = NULL, most_recent = FALSE)

search_spatial(wkt = NULL, location = NULL, distance = NULL,
  ids = NULL, most_recent = FALSE)

Arguments

dates
A vector of length 1 or 2, containing either ISO format date character strings ('yyyy-mm-dd') or POSIXt dates.

match_type
A character string (see Details).

most_recent
Logical indicating whether to restrict the API to returning only the most recent versions of the datasets found. By default all versions of matching dataset concepts are returned.

ids
A set of SAFE dataset record IDs to restrict a search. This will typically be a safe_record_set object returned by another search but can also be a vector of record ids in any of the formats accepted by validate_record_ids.

field_text
Text to search for within the data worksheet field name and description.

field_type
A data worksheet field type (see Links).

author
A character string used to search for datasets by author full (or partial) names.

taxon_name
The scientific name of a taxon to search for.

taxon_rank
A taxonomic rank to search for.

gbif_id
A GBIF taxonomic ID number.

text
Character string to look for within a SAFE dataset, worksheet, title, field description, and dataset keywords.

wkt
A well-known text geometry string, assumed to use latitude and longitude in WGS84 (EPSG:4326).

location
The name of a location in the SAFE gazetteer.

distance
A buffer distance for spatial searches, giving the distance in metres within which to match either location or wkt searches.

Value

An object of class safe_record_set of datasets that match the query.

Details

The API provides endpoints to search datasets by date extents, data worksheet fields, authors, taxa, free text and by spatial query. All of the functions accept the argument most_recent, which restricts the returned datasets to the most recent versions of each matching dataset concept. The functions can also be passed an existing safe_record_set object to search within the results of a previous search.

The match_type parameter specifies how to match date ranges and must be one of 'intersect' (default), 'contain', or 'within'. The 'contain' option returns datasets that span a date range, 'within' returns datasets that fall within the given range and 'intersect' selects datasets that overlap any part of the date range. Note that match_type is ignored when only a single date is provided.

Functions

search_dates: Search datasets by date extent

search_fields: Search data worksheet field metadata.

search_authors: Search by dataset author

search_taxa: Search by taxon name, rank or GBIF ID.

search_text: Search dataset, worksheet and field titles and descriptions

search_spatial: Search by spatial sampling area/named location.

Spatial searches

For spatial searches, users can select a location name from the SAFE gazetteer (see https://www.safeproject.net/info/gazetteer or load_gazetteer) or provide a WKT geometry. The sampling locations provided in each SAFE dataset are tested to see if they intersect the search geometry.

A buffer distance can aso be provided to extend the search around the query geometry. Note that although WKT geometries should be provided using WGS84 lat/long coordinates, since this is typical field GPS data, distances must be provided as metres and all proximity calculations take place in the UTM50N projected coordinate system.

The search_spatial function will not retrieve datasets that have not provided sampling locations or use newly defined locations that are missing coordinate information.

Links

SAFE API

https://www.safeproject.net/api

Worksheet field types

https://safe-dataset-checker.readthedocs.io/en/latest/data_format/data/#field-types

SAFE gazetteer

See load_gazetteer and https://www.safeproject.net/info/gazetteer

WKT

https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry

Examples

   search_dates('2014-06-12')

#> Search returned 31 records

#> Set includes 23 concept ids and 31 record ids: 
#>  - 18 open and most recent (*)
#>  - 7 open and outdated (o)
#>  - 6 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1198465 1237733         *
#> 2  ------- 1198466         o
#> 3  1198468 1237731         *
#> 4  ------- 1198469         o
#> 5  1198692 2599732         x
#> 6  ------- 1198693         x
#> 7  1198696 1198846         *
#> 8  ------- 1198697         o
#> 9  1198840 1237719         *
#> 10 ------- 1198841         o
#> 11 1216002 1304118         *
#> 12 ------- 1237657         o
#> 13 ------- 1216003         o
#> 14 1303011 1303012         *
#> 15 1303013 1303014         *
#> 16 1327556 1327557         *
#> 17 1344568 1344569         *
#> 18 1400563 1400564         x
#> 19 1441584 1441585         *
#> 20 1478523 1478524         x
#> 21 1478525 1478526         *
#> 22 1995247 1995439         *
#> 23 ------- 1995248         o
#> 24 3247454 3247455         *
#> 25 3247601 3247602         x
#> 26 3247630 3247631         x
#> 27 3247638 3247639         *
#> 28 3266770 3266771         *
#> 29 3266826 3266827         *
#> 30 3342494 3342495         *
#> 31 3354067 3354068         *

   search_dates(as.POSIXct(c('2014-06-12', '2015-06-11')))

#> Search returned 46 records

#> Set includes 35 concept ids and 46 record ids: 
#>  - 26 open and most recent (*)
#>  - 9 open and outdated (o)
#>  - 11 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1198465 1237733         *
#> 2  ------- 1198466         o
#> 3  1198468 1237731         *
#> 4  ------- 1198469         o
#> 5  1198471 1237732         *
#> 6  ------- 1198472         o
#> 7  1198692 2599732         x
#> 8  ------- 1198693         x
#> 9  1198696 1198846         *
#> 10 ------- 1198697         o
#> 11 1198840 1237719         *
#> 12 ------- 1198841         o
#> 13 1216002 1304118         *
#> 14 ------- 1237657         o
#> 15 ------- 1216003         o
#> 16 1303011 1303012         *
#> 17 1303013 1303014         *
#> 18 1303019 1303020         *
#> 19 1327556 1327557         *
#> 20 1344568 1344569         *
#> 21 1400563 1400564         x
#> 22 1441584 1441585         *
#> 23 1478523 1478524         x
#> 24 1478525 1478526         *
#> 25 1487594 1487595         *
#> 26 1995247 1995439         *
#> 27 ------- 1995248         o
#> 28 2020339 2020340         *
#> 29 2536268 2536269         x
#> 30 3081058 3247699         *
#> 31 ------- 3081059         o
#> 32 3247454 3247455         *
#> 33 3247464 3247465         x
#> 34 3247493 3247494         x
#> 35 3247601 3247602         x
#> 36 3247630 3247631         x
#> 37 3247638 3247639         *
#> 38 3251899 3258117         x
#> 39 ------- 3251900         x
#> 40 3265703 3265704         *
#> 41 3265745 3265746         *
#> 42 3265747 3265748         *
#> 43 3266770 3266771         *
#> 44 3266826 3266827         *
#> 45 3342494 3342495         *
#> 46 3354067 3354068         *

   search_dates(c('2014-06-12', '2015-06-11'), match_type = 'contain')

#> Search returned 5 records

#> Set includes 4 concept ids and 5 record ids: 
#>  - 3 open and most recent (*)
#>  - 1 open and outdated (o)
#>  - 1 under embargo or restricted (x)
#> 
#>   concept  record available
#> 1 2020339 2020340         *
#> 2 3081058 3247699         *
#> 3 ------- 3081059         o
#> 4 3247493 3247494         x
#> 5 3265703 3265704         *

   search_fields(field_text = 'temperature')

#> Search returned 7 records

#> Set includes 6 concept ids and 7 record ids: 
#>  - 3 open and most recent (*)
#>  - 0 open and outdated (o)
#>  - 4 under embargo or restricted (x)
#> 
#>   concept  record available
#> 1 3247484 3247485         x
#> 2 3247591 3247592         *
#> 3 3251886 3258079         x
#> 4 3251899 3258117         x
#> 5 ------- 3251900         x
#> 6 3266770 3266771         *
#> 7 3266821 3266822         *

   search_fields(field_type = 'numeric')

#> Search returned 26 records

#> Set includes 23 concept ids and 26 record ids: 
#>  - 13 open and most recent (*)
#>  - 0 open and outdated (o)
#>  - 13 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  3081058 3247699         *
#> 2  3237506 3237507         x
#> 3  3247454 3247455         *
#> 4  3247464 3247465         x
#> 5  3247484 3247485         x
#> 6  3247491 3247492         *
#> 7  3247493 3247494         x
#> 8  3247591 3247592         *
#> 9  3247601 3247602         x
#> 10 3247630 3247631         x
#> 11 3247638 3247639         *
#> 12 3251886 3258079         x
#> 13 ------- 3251887         x
#> 14 3251899 3258117         x
#> 15 ------- 3251900         x
#> 16 3251901 3258082         x
#> 17 ------- 3251902         x
#> 18 3265703 3265704         *
#> 19 3265711 3265712         x
#> 20 3265721 3265722         *
#> 21 3265745 3265746         *
#> 22 3265747 3265748         *
#> 23 3266770 3266771         *
#> 24 3266821 3266822         *
#> 25 3342494 3342495         *
#> 26 3354067 3354068         *

   search_fields(field_text = 'temperature', field_type = 'numeric')

#> Search returned 7 records

#> Set includes 6 concept ids and 7 record ids: 
#>  - 3 open and most recent (*)
#>  - 0 open and outdated (o)
#>  - 4 under embargo or restricted (x)
#> 
#>   concept  record available
#> 1 3247484 3247485         x
#> 2 3247591 3247592         *
#> 3 3251886 3258079         x
#> 4 3251899 3258117         x
#> 5 ------- 3251900         x
#> 6 3266770 3266771         *
#> 7 3266821 3266822         *

   search_authors('Ewers')

#> Search returned 56 records

#> Set includes 46 concept ids and 56 record ids: 
#>  - 40 open and most recent (*)
#>  - 9 open and outdated (o)
#>  - 7 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1198301 1198302         *
#> 2  1198460 1237736         *
#> 3  1198465 1237733         *
#> 4  1198468 1237731         *
#> 5  1198471 1237732         *
#> 6  1198476 1198477         *
#> 7  1198522 1237724         *
#> 8  1198560 1237722         *
#> 9  1198568 1304117         *
#> 10 ------- 1237721         o
#> 11 1198576 1237737         *
#> 12 ------- 1198577         o
#> 13 1198578 1237738         *
#> 14 ------- 1198579         o
#> 15 1198580 1237726         *
#> 16 1198582 1237725         *
#> 17 1198584 1198585         *
#> 18 1198586 1198587         *
#> 19 1198596 1237739         *
#> 20 1198692 2599732         x
#> 21 ------- 1198693         x
#> 22 1198694 1198695         *
#> 23 1198696 1198846         *
#> 24 ------- 1198697         o
#> 25 1198838 1198839         *
#> 26 1198840 1237719         *
#> 27 1198842 1237720         *
#> 28 1216002 1304118         *
#> 29 ------- 1237657         o
#> 30 ------- 1216003         o
#> 31 1216039 2537074         *
#> 32 ------- 1400565         o
#> 33 ------- 1216040         o
#> 34 1216043 1216044         *
#> 35 1219160 1219161         *
#> 36 1220269 1220270         *
#> 37 1228187 1228188         *
#> 38 1228189 1228190         *
#> 39 1237729 1237730         *
#> 40 1303011 1303012         *
#> 41 1303013 1303014         *
#> 42 1303017 1303018         *
#> 43 1303019 1303020         *
#> 44 1323503 1323504         *
#> 45 1344568 1344569         *
#> 46 1400563 1400564         x
#> 47 1441584 1441585         *
#> 48 2537072 2537073         x
#> 49 3081058 3247699         *
#> 50 ------- 3081059         o
#> 51 3237506 3237507         x
#> 52 3247484 3247485         x
#> 53 3265711 3265712         x
#> 54 3265721 3265722         *
#> 55 3266770 3266771         *
#> 56 3354067 3354068         *

   search_taxa(taxon_name = 'Formicidae')

#> Search returned 9 records

#> Set includes 9 concept ids and 9 record ids: 
#>  - 8 open and most recent (*)
#>  - 0 open and outdated (o)
#>  - 1 under embargo or restricted (x)
#> 
#>   concept  record available
#> 1 1198301 1198302         *
#> 2 1198471 1237732         *
#> 3 1198838 1198839         *
#> 4 1237729 1237730         *
#> 5 1400561 1400562         *
#> 6 1995247 1995439         *
#> 7 3247484 3247485         x
#> 8 3265745 3265746         *
#> 9 3354067 3354068         *

   search_taxa(gbif_id = 4342)

#> Search returned 9 records

#> Set includes 9 concept ids and 9 record ids: 
#>  - 8 open and most recent (*)
#>  - 0 open and outdated (o)
#>  - 1 under embargo or restricted (x)
#> 
#>   concept  record available
#> 1 1198301 1198302         *
#> 2 1198471 1237732         *
#> 3 1198838 1198839         *
#> 4 1237729 1237730         *
#> 5 1400561 1400562         *
#> 6 1995247 1995439         *
#> 7 3247484 3247485         x
#> 8 3265745 3265746         *
#> 9 3354067 3354068         *

   search_taxa(taxon_rank = 'family')

#> Search returned 46 records

#> Set includes 46 concept ids and 46 record ids: 
#>  - 36 open and most recent (*)
#>  - 0 open and outdated (o)
#>  - 10 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1198301 1198302         *
#> 2  1198460 1237736         *
#> 3  1198465 1237733         *
#> 4  1198468 1237731         *
#> 5  1198471 1237732         *
#> 6  1198474 1198475         *
#> 7  1198476 1198477         *
#> 8  1198522 1237724         *
#> 9  1198560 1237722         *
#> 10 1198568 1304117         *
#> 11 1198596 1237739         *
#> 12 1198692 2599732         x
#> 13 1198694 1198695         *
#> 14 1198696 1198846         *
#> 15 1198832 1198833         *
#> 16 1198838 1198839         *
#> 17 1198840 1237719         *
#> 18 1198842 1237720         *
#> 19 1216039 2537074         *
#> 20 1219160 1219161         *
#> 21 1237729 1237730         *
#> 22 1303009 1303010         *
#> 23 1303019 1303020         *
#> 24 1323503 1323504         *
#> 25 1327556 1327557         *
#> 26 1400561 1400562         *
#> 27 1400563 1400564         x
#> 28 1478523 1478524         x
#> 29 1478525 1478526         *
#> 30 1487594 1487595         *
#> 31 1995247 1995439         *
#> 32 2020339 2020340         *
#> 33 2024580 2536270         *
#> 34 2536268 2536269         x
#> 35 2579791 2579792         *
#> 36 3237506 3237507         x
#> 37 3247464 3247465         x
#> 38 3247484 3247485         x
#> 39 3247491 3247492         *
#> 40 3247493 3247494         x
#> 41 3247630 3247631         x
#> 42 3265703 3265704         *
#> 43 3265711 3265712         x
#> 44 3265745 3265746         *
#> 45 3342494 3342495         *
#> 46 3354067 3354068         *

   search_text('forest')

#> Search returned 20 records

#> Set includes 18 concept ids and 20 record ids: 
#>  - 10 open and most recent (*)
#>  - 0 open and outdated (o)
#>  - 10 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  3081058 3247699         *
#> 2  3237506 3237507         x
#> 3  3247454 3247455         *
#> 4  3247464 3247465         x
#> 5  3247484 3247485         x
#> 6  3247591 3247592         *
#> 7  3247601 3247602         x
#> 8  3247630 3247631         x
#> 9  3247638 3247639         *
#> 10 3251886 3258079         x
#> 11 ------- 3251887         x
#> 12 3251899 3258117         x
#> 13 ------- 3251900         x
#> 14 3265703 3265704         *
#> 15 3265711 3265712         x
#> 16 3265745 3265746         *
#> 17 3265747 3265748         *
#> 18 3266770 3266771         *
#> 19 3266821 3266822         *
#> 20 3266826 3266827         *

   search_text('ant')

#> Search returned 17 records

#> Set includes 17 concept ids and 17 record ids: 
#>  - 11 open and most recent (*)
#>  - 0 open and outdated (o)
#>  - 6 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  3081058 3247699         *
#> 2  3247454 3247455         *
#> 3  3247464 3247465         x
#> 4  3247484 3247485         x
#> 5  3247491 3247492         *
#> 6  3247493 3247494         x
#> 7  3247601 3247602         x
#> 8  3247630 3247631         x
#> 9  3247638 3247639         *
#> 10 3265703 3265704         *
#> 11 3265711 3265712         x
#> 12 3265745 3265746         *
#> 13 3265747 3265748         *
#> 14 3266770 3266771         *
#> 15 3266821 3266822         *
#> 16 3266826 3266827         *
#> 17 3354067 3354068         *

   search_spatial(wkt = 'Point(116.5 4.75)')

#> Search returned 0 records

#> Set includes 0 concept ids and 0 record ids: 
#>  - 0 open and most recent (*)
#>  - 0 open and outdated (o)
#>  - 0 under embargo or restricted (x)
#> 
#> [1] concept   record    available
#> <0 rows> (or 0-length row.names)

   search_spatial(wkt = 'Point(116.5 4.75)', distance = 100000)

#> Search returned 47 records

#> Set includes 40 concept ids and 47 record ids: 
#>  - 32 open and most recent (*)
#>  - 6 open and outdated (o)
#>  - 9 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1198460 1237736         *
#> 2  ------- 1198461         o
#> 3  1198476 1198477         *
#> 4  1198564 1198565         *
#> 5  1198576 1237737         *
#> 6  ------- 1198577         o
#> 7  1198578 1237738         *
#> 8  ------- 1198579         o
#> 9  1198580 1237726         *
#> 10 ------- 1198581         o
#> 11 1198582 1237725         *
#> 12 ------- 1198583         o
#> 13 1198584 1198585         *
#> 14 1198586 1198587         *
#> 15 1198692 2599732         x
#> 16 ------- 1198693         x
#> 17 1198696 1198846         *
#> 18 ------- 1198697         o
#> 19 1198832 1198833         *
#> 20 1198838 1198839         *
#> 21 1216043 1216044         *
#> 22 1220269 1220270         *
#> 23 1228187 1228188         *
#> 24 1228189 1228190         *
#> 25 1237729 1237730         *
#> 26 1303017 1303018         *
#> 27 1303019 1303020         *
#> 28 1323503 1323504         *
#> 29 1344568 1344569         *
#> 30 1400563 1400564         x
#> 31 1441584 1441585         *
#> 32 1478523 1478524         x
#> 33 2537072 2537073         x
#> 34 3247454 3247455         *
#> 35 3247484 3247485         x
#> 36 3247491 3247492         *
#> 37 3247493 3247494         x
#> 38 3247591 3247592         *
#> 39 3247601 3247602         x
#> 40 3247630 3247631         x
#> 41 3247638 3247639         *
#> 42 3265721 3265722         *
#> 43 3265745 3265746         *
#> 44 3265747 3265748         *
#> 45 3266770 3266771         *
#> 46 3342494 3342495         *
#> 47 3354067 3354068         *

   search_spatial(wkt = 'Polygon((110 0, 110 10,120 10,120 0,110 0))')

#> Search returned 74 records

#> Set includes 58 concept ids and 74 record ids: 
#>  - 48 open and most recent (*)
#>  - 15 open and outdated (o)
#>  - 11 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1198460 1237736         *
#> 2  ------- 1198461         o
#> 3  1198465 1237733         *
#> 4  ------- 1198466         o
#> 5  1198468 1237731         *
#> 6  ------- 1198469         o
#> 7  1198471 1237732         *
#> 8  ------- 1198472         o
#> 9  1198474 1198475         *
#> 10 1198476 1198477         *
#> 11 1198556 1198557         *
#> 12 1198560 1237722         *
#> 13 ------- 1198561         o
#> 14 1198564 1198565         *
#> 15 1198576 1237737         *
#> 16 ------- 1198577         o
#> 17 1198578 1237738         *
#> 18 ------- 1198579         o
#> 19 1198580 1237726         *
#> 20 ------- 1198581         o
#> 21 1198582 1237725         *
#> 22 ------- 1198583         o
#> 23 1198584 1198585         *
#> 24 1198586 1198587         *
#> 25 1198596 1237739         *
#> 26 ------- 1198597         o
#> 27 1198692 2599732         x
#> 28 ------- 1198693         x
#> 29 1198696 1198846         *
#> 30 ------- 1198697         o
#> 31 1198832 1198833         *
#> 32 1198838 1198839         *
#> 33 1198842 1237720         *
#> 34 ------- 1198843         o
#> 35 1216043 1216044         *
#> 36 1219160 1219161         *
#> 37 1220269 1220270         *
#> 38 1228187 1228188         *
#> 39 1228189 1228190         *
#> 40 1237729 1237730         *
#> 41 1303009 1303010         *
#> 42 1303017 1303018         *
#> 43 1303019 1303020         *
#> 44 1323503 1323504         *
#> 45 1327556 1327557         *
#> 46 1344568 1344569         *
#> 47 1400563 1400564         x
#> 48 1441584 1441585         *
#> 49 1478523 1478524         x
#> 50 1478525 1478526         *
#> 51 1487594 1487595         *
#> 52 1995247 1995439         *
#> 53 ------- 1995248         o
#> 54 2024580 2536270         *
#> 55 ------- 2024581         o
#> 56 2536268 2536269         x
#> 57 2537072 2537073         x
#> 58 3081058 3247699         *
#> 59 ------- 3081059         o
#> 60 3237506 3237507         x
#> 61 3247454 3247455         *
#> 62 3247484 3247485         x
#> 63 3247491 3247492         *
#> 64 3247493 3247494         x
#> 65 3247591 3247592         *
#> 66 3247601 3247602         x
#> 67 3247630 3247631         x
#> 68 3247638 3247639         *
#> 69 3265721 3265722         *
#> 70 3265745 3265746         *
#> 71 3265747 3265748         *
#> 72 3266770 3266771         *
#> 73 3342494 3342495         *
#> 74 3354067 3354068         *

   search_spatial(location = 'A_1')

#> Search returned 7 records

#> Set includes 6 concept ids and 7 record ids: 
#>  - 6 open and most recent (*)
#>  - 1 open and outdated (o)
#>  - 0 under embargo or restricted (x)
#> 
#>   concept  record available
#> 1 1198460 1237736         *
#> 2 ------- 1198461         o
#> 3 1198476 1198477         *
#> 4 1228187 1228188         *
#> 5 1323503 1323504         *
#> 6 1344568 1344569         *
#> 7 3354067 3354068         *

   search_spatial(location = 'A_1', distance = 2500)

#> Search returned 32 records

#> Set includes 26 concept ids and 32 record ids: 
#>  - 22 open and most recent (*)
#>  - 6 open and outdated (o)
#>  - 4 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1198460 1237736         *
#> 2  ------- 1198461         o
#> 3  1198468 1237731         *
#> 4  ------- 1198469         o
#> 5  1198476 1198477         *
#> 6  1198576 1237737         *
#> 7  ------- 1198577         o
#> 8  1198578 1237738         *
#> 9  ------- 1198579         o
#> 10 1198584 1198585         *
#> 11 1198586 1198587         *
#> 12 1228187 1228188         *
#> 13 1228189 1228190         *
#> 14 1237729 1237730         *
#> 15 1303009 1303010         *
#> 16 1303017 1303018         *
#> 17 1323503 1323504         *
#> 18 1344568 1344569         *
#> 19 1400563 1400564         x
#> 20 1441584 1441585         *
#> 21 1478525 1478526         *
#> 22 1487594 1487595         *
#> 23 1995247 1995439         *
#> 24 ------- 1995248         o
#> 25 2024580 2536270         *
#> 26 ------- 2024581         o
#> 27 2536268 2536269         x
#> 28 3247484 3247485         x
#> 29 3247491 3247492         *
#> 30 3247493 3247494         x
#> 31 3265721 3265722         *
#> 32 3354067 3354068         *