SAFE dataset search functions.

In addition to the datasets stored on Zenodo, the SAFE Project website provides an API to search dataset metadata in more depth. The search functions access this API and return safe_record_set objects identifying datasets that match a particular query.

search_dates(dates, match_type = "intersect", most_recent = FALSE, ids = NULL)

search_fields(
  field_text = NULL,
  field_type = NULL,
  ids = NULL,
  most_recent = FALSE
)

search_authors(author, ids = NULL, most_recent = FALSE)

search_taxa(
  taxon_name = NULL,
  taxon_rank = NULL,
  gbif_id = NULL,
  ids = NULL,
  most_recent = FALSE
)

search_text(text, ids = NULL, most_recent = FALSE)

search_spatial(
  wkt = NULL,
  location = NULL,
  distance = NULL,
  ids = NULL,
  most_recent = FALSE
)

Arguments

dates
A vector of length 1 or 2, containing either ISO format date character strings ("yyyy-mm-dd") or POSIXt dates.

match_type
A character string (see Details).

most_recent
Logical indicating whether to restrict the API to returning only the most recent versions of the datasets found. By default all versions of matching dataset concepts are returned.

ids
A set of SAFE dataset record IDs to restrict a search. This will typically be a safe_record_set object returned by another search but can also be a vector of record ids in any of the formats accepted by validate_record_ids.

field_text
Text to search for within the data worksheet field name and description.

field_type
A data worksheet field type (see Links).

author
A character string used to search for datasets by author full (or partial) names.

taxon_name
The scientific name of a taxon to search for.

taxon_rank
A taxonomic rank to search for.

gbif_id
A GBIF taxonomic ID number.

text
Character string to look for within a SAFE dataset, worksheet, title, field description, and dataset keywords.

wkt
A well-known text geometry string, assumed to use latitude and longitude in WGS84 (EPSG:4326).

location
The name of a location in the SAFE gazetteer.

distance
A buffer distance for spatial searches, giving the distance in metres within which to match either location or wkt searches.

Value

An object of class safe_record_set of datasets that match the query.

Details

The API provides endpoints to search datasets by date extents, data worksheet fields, authors, taxa, free text and by spatial query. All of the functions accept the argument most_recent, which restricts the returned datasets to the most recent versions of each matching dataset concept. The functions can also be passed an existing safe_record_set object to search within the results of a previous search.

The match_type parameter specifies how to match date ranges and must be one of "intersect" (default), "contain", or "within". The "contain" option returns datasets that span a date range, "within" returns datasets that fall within the given range and "intersect" selects datasets that overlap any part of the date range. Note that match_type is ignored when only a single date is provided.

Functions

search_dates: Search datasets by date extent

search_fields: Search data worksheet field metadata.

search_authors: Search by dataset author

search_taxa: Search by taxon name, rank or GBIF ID.

search_text: Search dataset, worksheet and field titles and descriptions

search_spatial: Search by spatial sampling area/named location.

Spatial searches

For spatial searches, users can select a location name from a SAFE data gazetteer (see e.g. https://www.safeproject.net/info/gazetteer or load_gazetteer) or provide a WKT geometry. The sampling locations provided in each SAFE dataset are tested to see if they intersect the search geometry.

A buffer distance can aso be provided to extend the search around the query geometry. Note that although WKT geometries should be provided using WGS84 lat/long coordinates, since this is typical field GPS data, distances must be provided as metres and all proximity calculations take place in the UTM50N projected coordinate system.

The search_spatial function will not retrieve datasets that have not provided sampling locations or use newly defined locations that are missing coordinate information.

Links

SAFE data API

e.g. https://www.safeproject.net/api

Worksheet field types

https://safedata-validator.readthedocs.io/en/latest/data_format/data.html#field-types

SAFE gazetteer

See load_gazetteer and e.g. https://www.safeproject.net/info/gazetteer

WKT

https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry

Examples

   # \donttest{
   search_dates("2014-06-12")

#> Search returned 62 records

#> Set includes 46 concept ids and 62 record ids: 
#>  - 39 open and most recent (*)
#>  - 13 open and outdated (o)
#>  - 10 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   629855 3490675         *
#> 2   629942 3492153         *
#> 3   630004 3490488         *
#> 4   630015 3492009         *
#> 5   630063 3697796         *
#> 6   630119 3492148         *
#> 7   630123 3491956         *
#> 8   630132 3490613         *
#> 9  ------- 3490564         o
#> 10  630161 3492064         *
#> 11  630228 3697804         *
#> 12  630241 3492126         *
#> 13  630259 3490687         *
#> 14  654585 3698297         *
#> 15 1198465 1237733         *
#> 16 ------- 1198466         o
#> 17 1198468 1237731         *
#> 18 ------- 1198469         o
#> 19 1198692 2599732         *
#> 20 ------- 1198693         o
#> 21 1198696 1198846         *
#> 22 ------- 1198697         o
#> 23 1198840 1237719         *
#> 24 ------- 1198841         o
#> 25 1216002 1304118         *
#> 26 ------- 1237657         o
#> 27 ------- 1216003         o
#> 28 1303011 1303012         *
#> 29 1303013 1303014         *
#> 30 1327556 1327557         *
#> 31 1344568 1344569         *
#> 32 1400563 1400564         *
#> 33 1441584 1441585         *
#> 34 1478523 3955050         x
#> 35 ------- 1478524         *
#> 36 1478525 4072959         x
#> 37 ------- 3982665         x
#> 38 ------- 1478526         *
#> 39 1995247 1995439         *
#> 40 ------- 1995248         o
#> 41 3247601 3247602         x
#> 42 3247630 3247631         x
#> 43 3247638 3247639         *
#> 44 3266770 3266771         x
#> 45 3266826 3266827         *
#> 46 3342494 3342495         *
#> 47 3354067 3354068         *
#> 48 3366103 3366104         *
#> 49 3402745 3402746         x
#> 50 3402747 3402748         x
#> 51 3402751 3402752         x
#> 52 3476357 3476390         *
#> 53 ------- 3476358         o
#> 54 3485120 3906082         *
#> 55 ------- 3833856         o
#> 56 ------- 3485121         o
#> 57 3780390 3780391         x
#> 58 3888374 3888375         *
#> 59 3901504 3971012         *
#> 60 ------- 3901505         o
#> 61 3908248 3908249         *
#> 62 3948442 3948443         *

   search_dates(as.POSIXct(c("2014-06-12", "2015-06-11")))

#> Search returned 89 records

#> Set includes 69 concept ids and 89 record ids: 
#>  - 54 open and most recent (*)
#>  - 17 open and outdated (o)
#>  - 18 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   629855 3490675         *
#> 2   629942 3492153         *
#> 3   630004 3490488         *
#> 4   630015 3492009         *
#> 5   630063 3697796         *
#> 6   630119 3492148         *
#> 7   630123 3491956         *
#> 8   630132 3490613         *
#> 9  ------- 3490564         o
#> 10  630161 3492064         *
#> 11  630228 3697804         *
#> 12  630241 3492126         *
#> 13  630259 3490687         *
#> 14  654585 3698297         *
#> 15 1198465 1237733         *
#> 16 ------- 1198466         o
#> 17 1198468 1237731         *
#> 18 ------- 1198469         o
#> 19 1198471 1237732         *
#> 20 ------- 1198472         o
#> 21 1198692 2599732         *
#> 22 ------- 1198693         o
#> 23 1198696 1198846         *
#> 24 ------- 1198697         o
#> 25 1198840 1237719         *
#> 26 ------- 1198841         o
#> 27 1216002 1304118         *
#> 28 ------- 1237657         o
#> 29 ------- 1216003         o
#> 30 1303011 1303012         *
#> 31 1303013 1303014         *
#> 32 1303019 1303020         *
#> 33 1327556 1327557         *
#> 34 1344568 1344569         *
#> 35 1400563 1400564         *
#> 36 1441584 1441585         *
#> 37 1478523 3955050         x
#> 38 ------- 1478524         *
#> 39 1478525 4072959         x
#> 40 ------- 3982665         x
#> 41 ------- 1478526         *
#> 42 1487594 1487595         *
#> 43 1995247 1995439         *
#> 44 ------- 1995248         o
#> 45 2020339 2020340         *
#> 46 2536268 3476542         *
#> 47 ------- 2536269         o
#> 48 3081058 3247699         *
#> 49 ------- 3081059         o
#> 50 3247464 3247465         x
#> 51 3247493 3247494         x
#> 52 3247601 3247602         x
#> 53 3247630 3247631         x
#> 54 3247638 3247639         *
#> 55 3251899 3258117         *
#> 56 ------- 3251900         o
#> 57 3265703 3265704         *
#> 58 3265745 3265746         *
#> 59 3265747 3265748         *
#> 60 3266770 3266771         x
#> 61 3266826 3266827         *
#> 62 3342494 3342495         *
#> 63 3354067 3354068         *
#> 64 3366103 3366104         *
#> 65 3402745 3402746         x
#> 66 3402747 3402748         x
#> 67 3402751 3402752         x
#> 68 3476357 3476390         *
#> 69 ------- 3476358         o
#> 70 3485120 3906082         *
#> 71 ------- 3833856         o
#> 72 ------- 3485121         o
#> 73 3698114 3698115         x
#> 74 3780390 3780391         x
#> 75 3888374 3888375         *
#> 76 3901504 3971012         *
#> 77 ------- 3901505         o
#> 78 3906117 3906118         x
#> 79 3908127 3908128         *
#> 80 3908248 3908249         *
#> 81 3929631 3929632         *
#> 82 3948442 3948443         *
#> 83 3973690 3973691         *
#> 84 3974970 3974971         x
#> 85 4010756 4010757         *
#> 86 4020696 4020697         *
#> 87 4139684 4139685         x
#> 88 4247168 4247169         x
#> 89 4275385 4275386         x

   search_dates(c("2014-06-12", "2015-06-11"), match_type = "contain")

#> Search returned 8 records

#> Set includes 7 concept ids and 8 record ids: 
#>  - 4 open and most recent (*)
#>  - 1 open and outdated (o)
#>  - 3 under embargo or restricted (x)
#> 
#>   concept  record available
#> 1 2020339 2020340         *
#> 2 3081058 3247699         *
#> 3 ------- 3081059         o
#> 4 3247493 3247494         x
#> 5 3265703 3265704         *
#> 6 4020696 4020697         *
#> 7 4139684 4139685         x
#> 8 4247168 4247169         x

   search_fields(field_text = "temperature")

#> Search returned 22 records

#> Set includes 18 concept ids and 22 record ids: 
#>  - 12 open and most recent (*)
#>  - 2 open and outdated (o)
#>  - 8 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1478525 4072959         x
#> 2  ------- 3982665         x
#> 3  3247484 3876227         *
#> 4  ------- 3247485         o
#> 5  3247591 3247592         *
#> 6  3251886 3258079         *
#> 7  3251899 3258117         *
#> 8  ------- 3251900         o
#> 9  3266770 3266771         x
#> 10 3266821 3266822         *
#> 11 3475407 3475408         *
#> 12 3479170 3479171         *
#> 13 3698114 3698115         x
#> 14 3888374 3888375         *
#> 15 3908146 3929764         x
#> 16 ------- 3908147         x
#> 17 3926373 3926374         *
#> 18 3948442 3948443         *
#> 19 3952831 3952832         *
#> 20 3994259 3994260         x
#> 21 4000206 4000207         *
#> 22 4095373 4095374         x

   search_fields(field_type = "numeric")

#> Search returned 84 records

#> Set includes 69 concept ids and 84 record ids: 
#>  - 48 open and most recent (*)
#>  - 11 open and outdated (o)
#>  - 25 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   629855 3490675         *
#> 2   629942 3492153         *
#> 3   630015 3492009         *
#> 4   630119 3492148         *
#> 5   630132 3490613         *
#> 6   630161 3492064         *
#> 7   630241 3492126         *
#> 8  1198564 3857210         *
#> 9  1478523 3955050         x
#> 10 1478525 4072959         x
#> 11 ------- 3982665         x
#> 12 2536268 3476542         *
#> 13 2537072 3901735         x
#> 14 3081058 3247699         *
#> 15 3237506 3237507         x
#> 16 3247464 3247465         x
#> 17 3247484 3876227         *
#> 18 ------- 3247485         o
#> 19 3247491 3247492         *
#> 20 3247493 3247494         x
#> 21 3247591 3247592         *
#> 22 3247601 3247602         x
#> 23 3247630 3247631         x
#> 24 3247638 3247639         *
#> 25 3251886 3258079         *
#> 26 ------- 3251887         o
#> 27 3251899 3258117         *
#> 28 ------- 3251900         o
#> 29 3251901 3897394         *
#> 30 ------- 3614876         o
#> 31 ------- 3258082         o
#> 32 ------- 3251902         o
#> 33 3265703 3265704         *
#> 34 3265711 3997172         x
#> 35 ------- 3742834         x
#> 36 ------- 3265712         x
#> 37 3265721 3265722         *
#> 38 3265745 3265746         *
#> 39 3265747 3265748         *
#> 40 3266770 3266771         x
#> 41 3266821 3266822         *
#> 42 3342494 3342495         *
#> 43 3354067 3354068         *
#> 44 3366103 3366104         *
#> 45 3402745 3402746         x
#> 46 3475405 3475406         *
#> 47 3475407 3475408         *
#> 48 3476357 3476390         *
#> 49 ------- 3476358         o
#> 50 3479170 3479171         *
#> 51 3485085 3485086         x
#> 52 3485120 3906082         *
#> 53 ------- 3833856         o
#> 54 ------- 3485121         o
#> 55 3698114 3698115         x
#> 56 3780390 3780391         x
#> 57 3888374 3888375         *
#> 58 3897376 3897377         x
#> 59 3901504 3971012         *
#> 60 ------- 3901505         o
#> 61 3906117 3906118         x
#> 62 3908117 3908118         *
#> 63 3908127 3908128         *
#> 64 3908146 3929764         x
#> 65 ------- 3908147         x
#> 66 3908248 3908249         *
#> 67 3926373 3926374         *
#> 68 3929631 3929632         *
#> 69 3948442 3948443         *
#> 70 3952831 3952832         *
#> 71 3966236 3966277         *
#> 72 ------- 3966237         o
#> 73 3973550 3973551         *
#> 74 3973690 3973691         *
#> 75 3974970 3974971         x
#> 76 3975972 3975973         *
#> 77 3976020 3976021         *
#> 78 3979295 3979296         x
#> 79 3981221 3981222         *
#> 80 3994259 3994260         x
#> 81 4000206 4000207         *
#> 82 4010756 4010757         *
#> 83 4020696 4020697         *
#> 84 4095373 4095374         x

   search_fields(field_text = "temperature", field_type = "numeric")

#> Search returned 21 records

#> Set includes 17 concept ids and 21 record ids: 
#>  - 11 open and most recent (*)
#>  - 2 open and outdated (o)
#>  - 8 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1478525 4072959         x
#> 2  ------- 3982665         x
#> 3  3247484 3876227         *
#> 4  ------- 3247485         o
#> 5  3247591 3247592         *
#> 6  3251886 3258079         *
#> 7  3251899 3258117         *
#> 8  ------- 3251900         o
#> 9  3266770 3266771         x
#> 10 3266821 3266822         *
#> 11 3475407 3475408         *
#> 12 3479170 3479171         *
#> 13 3698114 3698115         x
#> 14 3888374 3888375         *
#> 15 3908146 3929764         x
#> 16 ------- 3908147         x
#> 17 3926373 3926374         *
#> 18 3952831 3952832         *
#> 19 3994259 3994260         x
#> 20 4000206 4000207         *
#> 21 4095373 4095374         x

   search_authors("Ewers")

#> Search returned 81 records

#> Set includes 63 concept ids and 81 record ids: 
#>  - 53 open and most recent (*)
#>  - 13 open and outdated (o)
#>  - 15 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   629855 3490675         *
#> 2   629942 3492153         *
#> 3   654585 3698297         *
#> 4  1198301 1198302         *
#> 5  1198460 1237736         *
#> 6  1198465 1237733         *
#> 7  1198468 1237731         *
#> 8  1198471 1237732         *
#> 9  1198476 1198477         *
#> 10 1198522 1237724         *
#> 11 1198560 1237722         *
#> 12 1198568 1304117         *
#> 13 ------- 1237721         o
#> 14 1198576 1237737         *
#> 15 ------- 1198577         o
#> 16 1198578 1237738         *
#> 17 ------- 1198579         o
#> 18 1198580 1237726         *
#> 19 1198582 1237725         *
#> 20 1198584 1198585         *
#> 21 1198586 1198587         *
#> 22 1198596 1237739         *
#> 23 1198692 2599732         *
#> 24 ------- 1198693         o
#> 25 1198694 1198695         *
#> 26 1198696 1198846         *
#> 27 ------- 1198697         o
#> 28 1198838 1198839         *
#> 29 1198840 1237719         *
#> 30 1198842 1237720         *
#> 31 1216002 1304118         *
#> 32 ------- 1237657         o
#> 33 ------- 1216003         o
#> 34 1216039 2537074         *
#> 35 ------- 1400565         o
#> 36 ------- 1216040         o
#> 37 1216043 1216044         *
#> 38 1219160 1219161         *
#> 39 1220269 1220270         *
#> 40 1228187 1228188         *
#> 41 1228189 1228190         *
#> 42 1237729 1237730         x
#> 43 1303011 1303012         *
#> 44 1303013 1303014         *
#> 45 1303017 1303018         *
#> 46 1303019 1303020         *
#> 47 1323503 1323504         *
#> 48 1344568 1344569         *
#> 49 1400563 1400564         *
#> 50 1441584 1441585         *
#> 51 1478523 3955050         x
#> 52 1478525 4072959         x
#> 53 ------- 3982665         x
#> 54 2537072 3901735         x
#> 55 ------- 2537073         x
#> 56 3081058 3247699         *
#> 57 ------- 3081059         o
#> 58 3237506 3237507         x
#> 59 3247484 3876227         *
#> 60 ------- 3247485         o
#> 61 3265711 3997172         x
#> 62 ------- 3742834         x
#> 63 ------- 3265712         x
#> 64 3265721 3265722         *
#> 65 3266770 3266771         x
#> 66 3354067 3354068         *
#> 67 3476357 3476390         *
#> 68 ------- 3476358         o
#> 69 3479170 3479171         *
#> 70 3888374 3888375         *
#> 71 3908117 3908118         *
#> 72 3908146 3929764         x
#> 73 ------- 3908147         x
#> 74 3966236 3966277         *
#> 75 ------- 3966237         o
#> 76 3973550 3973551         *
#> 77 3973690 3973691         *
#> 78 3974970 3974971         x
#> 79 3975972 3975973         *
#> 80 3981221 3981222         *
#> 81 3994259 3994260         x

   search_taxa(taxon_name = "Formicidae")

#> Search returned 13 records

#> Set includes 12 concept ids and 13 record ids: 
#>  - 9 open and most recent (*)
#>  - 1 open and outdated (o)
#>  - 3 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1198301 1198302         *
#> 2  1198471 1237732         *
#> 3  1198838 1198839         *
#> 4  1237729 1237730         x
#> 5  1400561 1400562         *
#> 6  1995247 1995439         *
#> 7  2537072 3901735         x
#> 8  3247484 3876227         *
#> 9  ------- 3247485         o
#> 10 3265745 3265746         *
#> 11 3354067 3354068         *
#> 12 3975972 3975973         *
#> 13 3979295 3979296         x

   search_taxa(gbif_id = 4342)

#> Search returned 13 records

#> Set includes 12 concept ids and 13 record ids: 
#>  - 9 open and most recent (*)
#>  - 1 open and outdated (o)
#>  - 3 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1198301 1198302         *
#> 2  1198471 1237732         *
#> 3  1198838 1198839         *
#> 4  1237729 1237730         x
#> 5  1400561 1400562         *
#> 6  1995247 1995439         *
#> 7  2537072 3901735         x
#> 8  3247484 3876227         *
#> 9  ------- 3247485         o
#> 10 3265745 3265746         *
#> 11 3354067 3354068         *
#> 12 3975972 3975973         *
#> 13 3979295 3979296         x

   search_taxa(taxon_rank = "family")

#> Search returned 90 records

#> Set includes 80 concept ids and 90 record ids: 
#>  - 57 open and most recent (*)
#>  - 4 open and outdated (o)
#>  - 29 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1  1198301 1198302         *
#> 2  1198460 1237736         *
#> 3  1198465 1237733         *
#> 4  1198468 1237731         *
#> 5  1198471 1237732         *
#> 6  1198474 1198475         *
#> 7  1198476 1198477         *
#> 8  1198522 1237724         *
#> 9  1198560 1237722         *
#> 10 1198568 1304117         *
#> 11 1198596 1237739         *
#> 12 1198692 2599732         *
#> 13 1198694 1198695         *
#> 14 1198696 1198846         *
#> 15 1198832 1198833         *
#> 16 1198838 1198839         *
#> 17 1198840 1237719         *
#> 18 1198842 1237720         *
#> 19 1216039 2537074         *
#> 20 1219160 1219161         *
#> 21 1237729 1237730         x
#> 22 1303009 1303010         *
#> 23 1303019 1303020         *
#> 24 1323503 1323504         *
#> 25 1327556 1327557         *
#> 26 1400561 1400562         *
#> 27 1400563 1400564         *
#> 28 1478523 3955050         x
#> 29 ------- 1478524         *
#> 30 1478525 4072959         x
#> 31 ------- 3982665         x
#> 32 ------- 1478526         *
#> 33 1487594 1487595         *
#> 34 1995247 1995439         *
#> 35 2020339 2020340         *
#> 36 2024580 2536270         *
#> 37 2536268 3476542         *
#> 38 ------- 2536269         o
#> 39 2537072 3901735         x
#> 40 2579791 2579792         *
#> 41 3237506 3237507         x
#> 42 3247464 3247465         x
#> 43 3247484 3876227         *
#> 44 ------- 3247485         o
#> 45 3247491 3247492         *
#> 46 3247493 3247494         x
#> 47 3247630 3247631         x
#> 48 3265703 3265704         *
#> 49 3265711 3997172         x
#> 50 ------- 3742834         x
#> 51 ------- 3265712         x
#> 52 3265745 3265746         *
#> 53 3342494 3342495         *
#> 54 3354067 3354068         *
#> 55 3366103 3366104         *
#> 56 3402747 3402748         x
#> 57 3402751 3402752         x
#> 58 3475405 3475406         *
#> 59 3475407 3475408         *
#> 60 3485085 3485086         x
#> 61 3780390 3780391         x
#> 62 3832075 3832076         x
#> 63 3897376 3897377         x
#> 64 3901504 3971012         *
#> 65 ------- 3901505         o
#> 66 3906117 3906118         x
#> 67 3906440 3906441         x
#> 68 3908127 3908128         *
#> 69 3908146 3929764         x
#> 70 ------- 3908147         x
#> 71 3908248 3908249         *
#> 72 3926373 3926374         *
#> 73 3948442 3948443         *
#> 74 3952831 3952832         *
#> 75 3966236 3966277         *
#> 76 ------- 3966237         o
#> 77 3973550 3973551         *
#> 78 3973690 3973691         *
#> 79 3974970 3974971         x
#> 80 3975204 3975205         *
#> 81 3975972 3975973         *
#> 82 3976020 3976021         *
#> 83 3979295 3979296         x
#> 84 3981221 3981222         *
#> 85 3994259 3994260         x
#> 86 4010756 4010757         *
#> 87 4095373 4095374         x
#> 88 4139684 4139685         x
#> 89 4247168 4247169         x
#> 90 4275385 4275386         x

   search_text("forest")

#> Search returned 70 records

#> Set includes 58 concept ids and 70 record ids: 
#>  - 35 open and most recent (*)
#>  - 8 open and outdated (o)
#>  - 27 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   629942 3492153         *
#> 2   630132 3490613         *
#> 3   630161 3492064         *
#> 4  1478523 3955050         x
#> 5  1478525 4072959         x
#> 6  ------- 3982665         x
#> 7  2536268 3476542         *
#> 8  2537072 3901735         x
#> 9  3081058 3247699         *
#> 10 3237506 3237507         x
#> 11 3247464 3247465         x
#> 12 3247484 3876227         *
#> 13 ------- 3247485         o
#> 14 3247591 3247592         *
#> 15 3247601 3247602         x
#> 16 3247630 3247631         x
#> 17 3247638 3247639         *
#> 18 3251886 3258079         *
#> 19 ------- 3251887         o
#> 20 3251899 3258117         *
#> 21 ------- 3251900         o
#> 22 3265703 3265704         *
#> 23 3265711 3997172         x
#> 24 ------- 3742834         x
#> 25 ------- 3265712         x
#> 26 3265745 3265746         *
#> 27 3265747 3265748         *
#> 28 3266770 3266771         x
#> 29 3266821 3266822         *
#> 30 3266826 3266827         *
#> 31 3366103 3366104         *
#> 32 3402747 3402748         x
#> 33 3402751 3402752         x
#> 34 3475405 3475406         *
#> 35 3475407 3475408         *
#> 36 3476357 3476390         *
#> 37 ------- 3476358         o
#> 38 3485085 3485086         x
#> 39 3485120 3906082         *
#> 40 ------- 3833856         o
#> 41 ------- 3485121         o
#> 42 3698114 3698115         x
#> 43 3780390 3780391         x
#> 44 3897376 3897377         x
#> 45 3901504 3971012         *
#> 46 ------- 3901505         o
#> 47 3908117 3908118         *
#> 48 3908127 3908128         *
#> 49 3908146 3929764         x
#> 50 ------- 3908147         x
#> 51 3926373 3926374         *
#> 52 3929631 3929632         *
#> 53 3948442 3948443         *
#> 54 3952831 3952832         *
#> 55 3966236 3966277         *
#> 56 ------- 3966237         o
#> 57 3973690 3973691         *
#> 58 3974970 3974971         x
#> 59 3975204 3975205         *
#> 60 3976020 3976021         *
#> 61 3979295 3979296         x
#> 62 3981221 3981222         *
#> 63 3994259 3994260         x
#> 64 4000206 4000207         *
#> 65 4010756 4010757         *
#> 66 4020696 4020697         *
#> 67 4095373 4095374         x
#> 68 4139684 4139685         x
#> 69 4247168 4247169         x
#> 70 4275385 4275386         x

   search_text("ant")

#> Search returned 60 records

#> Set includes 53 concept ids and 60 record ids: 
#>  - 31 open and most recent (*)
#>  - 3 open and outdated (o)
#>  - 26 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   630119 3492148         *
#> 2   630161 3492064         *
#> 3  1478523 3955050         x
#> 4  1478525 4072959         x
#> 5  ------- 3982665         x
#> 6  2536268 3476542         *
#> 7  2537072 3901735         x
#> 8  3081058 3247699         *
#> 9  3247464 3247465         x
#> 10 3247484 3876227         *
#> 11 ------- 3247485         o
#> 12 3247491 3247492         *
#> 13 3247493 3247494         x
#> 14 3247601 3247602         x
#> 15 3247630 3247631         x
#> 16 3247638 3247639         *
#> 17 3265703 3265704         *
#> 18 3265711 3997172         x
#> 19 ------- 3742834         x
#> 20 ------- 3265712         x
#> 21 3265745 3265746         *
#> 22 3265747 3265748         *
#> 23 3266770 3266771         x
#> 24 3266821 3266822         *
#> 25 3266826 3266827         *
#> 26 3354067 3354068         *
#> 27 3402745 3402746         x
#> 28 3402751 3402752         x
#> 29 3475405 3475406         *
#> 30 3475407 3475408         *
#> 31 3485085 3485086         x
#> 32 3780390 3780391         x
#> 33 3832075 3832076         x
#> 34 3888374 3888375         *
#> 35 3897376 3897377         x
#> 36 3901504 3971012         *
#> 37 ------- 3901505         o
#> 38 3906117 3906118         x
#> 39 3906440 3906441         x
#> 40 3908117 3908118         *
#> 41 3908127 3908128         *
#> 42 3908146 3929764         x
#> 43 ------- 3908147         x
#> 44 3926373 3926374         *
#> 45 3929631 3929632         *
#> 46 3948442 3948443         *
#> 47 3952831 3952832         *
#> 48 3966236 3966277         *
#> 49 ------- 3966237         o
#> 50 3973550 3973551         *
#> 51 3975204 3975205         *
#> 52 3975972 3975973         *
#> 53 3976020 3976021         *
#> 54 3979295 3979296         x
#> 55 3981221 3981222         *
#> 56 3994259 3994260         x
#> 57 4000206 4000207         *
#> 58 4020696 4020697         *
#> 59 4095373 4095374         x
#> 60 4139684 4139685         x

   search_spatial(wkt = "Point(116.5 4.75)")

#> Search returned 0 records

#> Set includes 0 concept ids and 0 record ids: 
#>  - 0 open and most recent (*)
#>  - 0 open and outdated (o)
#>  - 0 under embargo or restricted (x)
#> 
#> [1] concept   record    available
#> <0 rows> (or 0-length row.names)

   search_spatial(wkt = "Point(116.5 4.75)", distance = 100000)

#> Search returned 65 records

#> Set includes 53 concept ids and 65 record ids: 
#>  - 40 open and most recent (*)
#>  - 10 open and outdated (o)
#>  - 15 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   630015 3492009         *
#> 2   630132 3490613         *
#> 3  ------- 3490564         o
#> 4  1198460 1237736         *
#> 5  ------- 1198461         o
#> 6  1198476 1198477         *
#> 7  1198564 3857210         *
#> 8  ------- 1198565         o
#> 9  1198576 1237737         *
#> 10 ------- 1198577         o
#> 11 1198578 1237738         *
#> 12 ------- 1198579         o
#> 13 1198580 1237726         *
#> 14 ------- 1198581         o
#> 15 1198582 1237725         *
#> 16 ------- 1198583         o
#> 17 1198584 1198585         *
#> 18 1198586 1198587         *
#> 19 1198692 2599732         *
#> 20 ------- 1198693         o
#> 21 1198696 1198846         *
#> 22 ------- 1198697         o
#> 23 1198832 1198833         *
#> 24 1198838 1198839         *
#> 25 1216043 1216044         *
#> 26 1220269 1220270         *
#> 27 1228187 1228188         *
#> 28 1228189 1228190         *
#> 29 1237729 1237730         x
#> 30 1303017 1303018         *
#> 31 1303019 1303020         *
#> 32 1323503 1323504         *
#> 33 1344568 1344569         *
#> 34 1400563 1400564         *
#> 35 1441584 1441585         *
#> 36 1478523 3955050         x
#> 37 ------- 1478524         *
#> 38 2537072 3901735         x
#> 39 ------- 2537073         x
#> 40 3247484 3876227         *
#> 41 ------- 3247485         o
#> 42 3247491 3247492         *
#> 43 3247493 3247494         x
#> 44 3247591 3247592         *
#> 45 3247601 3247602         x
#> 46 3247630 3247631         x
#> 47 3247638 3247639         *
#> 48 3265721 3265722         *
#> 49 3265745 3265746         *
#> 50 3265747 3265748         *
#> 51 3266770 3266771         x
#> 52 3342494 3342495         *
#> 53 3354067 3354068         *
#> 54 3402747 3402748         x
#> 55 3402751 3402752         x
#> 56 3780390 3780391         x
#> 57 3906440 3906441         x
#> 58 3908117 3908118         *
#> 59 3908127 3908128         *
#> 60 3952831 3952832         *
#> 61 3973690 3973691         *
#> 62 3975204 3975205         *
#> 63 4139684 4139685         x
#> 64 4247168 4247169         x
#> 65 4275385 4275386         x

   search_spatial(wkt = "Polygon((110 0, 110 10,120 10,120 0,110 0))")

#> Search returned 119 records

#> Set includes 91 concept ids and 119 record ids: 
#>  - 69 open and most recent (*)
#>  - 23 open and outdated (o)
#>  - 27 under embargo or restricted (x)
#> 
#>     concept  record available
#> 1    629855 3490675         *
#> 2    630015 3492009         *
#> 3    630132 3490613         *
#> 4   ------- 3490564         o
#> 5   1198460 1237736         *
#> 6   ------- 1198461         o
#> 7   1198465 1237733         *
#> 8   ------- 1198466         o
#> 9   1198468 1237731         *
#> 10  ------- 1198469         o
#> 11  1198471 1237732         *
#> 12  ------- 1198472         o
#> 13  1198474 1198475         *
#> 14  1198476 1198477         *
#> 15  1198556 1198557         *
#> 16  1198560 1237722         *
#> 17  ------- 1198561         o
#> 18  1198564 3857210         *
#> 19  ------- 1198565         o
#> 20  1198576 1237737         *
#> 21  ------- 1198577         o
#> 22  1198578 1237738         *
#> 23  ------- 1198579         o
#> 24  1198580 1237726         *
#> 25  ------- 1198581         o
#> 26  1198582 1237725         *
#> 27  ------- 1198583         o
#> 28  1198584 1198585         *
#> 29  1198586 1198587         *
#> 30  1198596 1237739         *
#> 31  ------- 1198597         o
#> 32  1198692 2599732         *
#> 33  ------- 1198693         o
#> 34  1198696 1198846         *
#> 35  ------- 1198697         o
#> 36  1198832 1198833         *
#> 37  1198838 1198839         *
#> 38  1198842 1237720         *
#> 39  ------- 1198843         o
#> 40  1216043 1216044         *
#> 41  1219160 1219161         *
#> 42  1220269 1220270         *
#> 43  1228187 1228188         *
#> 44  1228189 1228190         *
#> 45  1237729 1237730         x
#> 46  1303009 1303010         *
#> 47  1303017 1303018         *
#> 48  1303019 1303020         *
#> 49  1323503 1323504         *
#> 50  1327556 1327557         *
#> 51  1344568 1344569         *
#> 52  1400563 1400564         *
#> 53  1441584 1441585         *
#> 54  1478523 3955050         x
#> 55  ------- 1478524         *
#> 56  1478525 4072959         x
#> 57  ------- 3982665         x
#> 58  ------- 1478526         *
#> 59  1995247 1995439         *
#> 60  ------- 1995248         o
#> 61  2024580 2536270         *
#> 62  ------- 2024581         o
#> 63  2536268 3476542         *
#> 64  ------- 2536269         o
#> 65  2537072 3901735         x
#> 66  ------- 2537073         x
#> 67  3081058 3247699         *
#> 68  ------- 3081059         o
#> 69  3237506 3237507         x
#> 70  3247484 3876227         *
#> 71  ------- 3247485         o
#> 72  3247491 3247492         *
#> 73  3247493 3247494         x
#> 74  3247591 3247592         *
#> 75  3247601 3247602         x
#> 76  3247630 3247631         x
#> 77  3247638 3247639         *
#> 78  3265721 3265722         *
#> 79  3265745 3265746         *
#> 80  3265747 3265748         *
#> 81  3266770 3266771         x
#> 82  3342494 3342495         *
#> 83  3354067 3354068         *
#> 84  3402747 3402748         x
#> 85  3402751 3402752         x
#> 86  3475407 3475408         *
#> 87  3476357 3476390         *
#> 88  ------- 3476358         o
#> 89  3479170 3479171         *
#> 90  3485085 3485086         x
#> 91  3780390 3780391         x
#> 92  3832075 3832076         x
#> 93  3897376 3897377         x
#> 94  3901504 3971012         *
#> 95  ------- 3901505         o
#> 96  3906117 3906118         x
#> 97  3906440 3906441         x
#> 98  3908117 3908118         *
#> 99  3908127 3908128         *
#> 100 3908146 3929764         x
#> 101 ------- 3908147         x
#> 102 3926373 3926374         *
#> 103 3929631 3929632         *
#> 104 3948442 3948443         *
#> 105 3952831 3952832         *
#> 106 3966236 3966277         *
#> 107 ------- 3966237         o
#> 108 3973550 3973551         *
#> 109 3973690 3973691         *
#> 110 3974970 3974971         x
#> 111 3975204 3975205         *
#> 112 3975972 3975973         *
#> 113 3976020 3976021         *
#> 114 3981221 3981222         *
#> 115 3994259 3994260         x
#> 116 4095373 4095374         x
#> 117 4139684 4139685         x
#> 118 4247168 4247169         x
#> 119 4275385 4275386         x

   search_spatial(location = "A_1")

#> Search returned 9 records

#> Set includes 7 concept ids and 9 record ids: 
#>  - 7 open and most recent (*)
#>  - 2 open and outdated (o)
#>  - 0 under embargo or restricted (x)
#> 
#>   concept  record available
#> 1 1198460 1237736         *
#> 2 ------- 1198461         o
#> 3 1198476 1198477         *
#> 4 1228187 1228188         *
#> 5 1323503 1323504         *
#> 6 1344568 1344569         *
#> 7 3354067 3354068         *
#> 8 3966236 3966277         *
#> 9 ------- 3966237         o

   search_spatial(location = "A_1", distance = 2500)

#> Search returned 54 records

#> Set includes 41 concept ids and 54 record ids: 
#>  - 32 open and most recent (*)
#>  - 11 open and outdated (o)
#>  - 11 under embargo or restricted (x)
#> 
#>    concept  record available
#> 1   629855 3490675         *
#> 2  1198460 1237736         *
#> 3  ------- 1198461         o
#> 4  1198468 1237731         *
#> 5  ------- 1198469         o
#> 6  1198476 1198477         *
#> 7  1198576 1237737         *
#> 8  ------- 1198577         o
#> 9  1198578 1237738         *
#> 10 ------- 1198579         o
#> 11 1198584 1198585         *
#> 12 1198586 1198587         *
#> 13 1228187 1228188         *
#> 14 1228189 1228190         *
#> 15 1237729 1237730         x
#> 16 1303009 1303010         *
#> 17 1303017 1303018         *
#> 18 1323503 1323504         *
#> 19 1344568 1344569         *
#> 20 1400563 1400564         *
#> 21 1441584 1441585         *
#> 22 1478525 4072959         x
#> 23 ------- 3982665         x
#> 24 ------- 1478526         *
#> 25 1995247 1995439         *
#> 26 ------- 1995248         o
#> 27 2024580 2536270         *
#> 28 ------- 2024581         o
#> 29 2536268 3476542         *
#> 30 ------- 2536269         o
#> 31 3247484 3876227         *
#> 32 ------- 3247485         o
#> 33 3247491 3247492         *
#> 34 3247493 3247494         x
#> 35 3265721 3265722         *
#> 36 3354067 3354068         *
#> 37 3476357 3476390         *
#> 38 ------- 3476358         o
#> 39 3485085 3485086         x
#> 40 3832075 3832076         x
#> 41 3897376 3897377         x
#> 42 3901504 3971012         *
#> 43 ------- 3901505         o
#> 44 3906117 3906118         x
#> 45 3908127 3908128         *
#> 46 3926373 3926374         *
#> 47 3966236 3966277         *
#> 48 ------- 3966237         o
#> 49 3973550 3973551         *
#> 50 3974970 3974971         x
#> 51 3975204 3975205         *
#> 52 3976020 3976021         *
#> 53 4139684 4139685         x
#> 54 4247168 4247169         x

   # combining searches using logical operators
   fish <- search_taxa('Actinopterygii')

#> Search returned 6 records

   odonates <- search_taxa("Odonata")

#> Search returned 3 records

   ewers <- search_authors("Ewers")

#> Search returned 81 records

   aquatic <- fish | odonates
   aquatic_ewers <- aquatic & ewers
   all_in_one <- (fish | odonates) & ewers
   # }